% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-aaft.R
\name{aaft_surrogate}
\alias{aaft_surrogate}
\title{Generate an amplitude adjusted Fourier transform surrogate according
to Theiler. The surrogate series preserves the amplitude distribution
(histogram) of the original time series, and are constrained to the
values of the original time series.}
\usage{
aaft_surrogate(series)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}
}
\description{
Generate an amplitude adjusted Fourier transform surrogate according
to Theiler. The surrogate series preserves the amplitude distribution
(histogram) of the original time series, and are constrained to the
values of the original time series.
}
