% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_forecast.R
\name{acum}
\alias{acum}
\title{Transform variables in percentage in variables in accumulate percentage}
\usage{
acum(data, n, exclude)
}
\arguments{
\item{data}{A tbl with variables in columns}

\item{n}{Number of periods to accumulate}

\item{exclude}{Variables to left aside}

}

\value{
An object of class tibble (tbl) with variables in accumulated percentage.
}
\description{
Given a tibble (or data frame) with variables in percentage the function creates a tibble with the variables in accumulated percentage.
}

\examples{
\dontrun{
data <- tibble(Date = seq.Date(from = as.Date("2012-01-01"),
                               to = as.Date("2018-06-01"),
                               by = "month"),
               V1 = rnorm(mean = 0, sd = 1, n = 78),
               V2 = rnorm(mean = 1, sd = 2, n = 78),
               V3 = rnorm(mean = 5, sd = 1, n = 78))

data_acum <- acum(data, n = 12)
              }
}
\author{
J. Renato Leripio
}
