% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_forecast.R
\name{acum_i}
\alias{acum_i}
\title{Transform variables in index form into variables in accumulate percentage in n periods relative to the previous n periods}
\usage{
acum_i(data, n, exclude)
}
\arguments{
\item{data}{A tbl with variables in columns}

\item{n}{Number of periods to accumulate}

\item{exclude}{Variables to left aside}

}

\value{
An object of class tibble (tbl) with variables in accumulated percentage.
}
\description{
Given a tibble (or data frame) with variables in index form the function creates a tibble with the variables in accumulated percentage in n periods relative to the previous n periods.
}

\examples{
\dontrun{

library(tidyverse)
library(tstools)

data <- tibble(Date = seq.Date(from = as.Date("2012-01-01"),
                               to = as.Date("2018-06-01"),
                               by = "month"),

               Index = rep(100, 78) + cumsum(rnorm(mean = 0, sd = 1, n = 78)))



data_acum <- acum_i(data, n = 12, "Date")
              }
}
\author{
J. Renato Leripio
}
