% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_add_grouping_column.R
\name{add_grouping_column}
\alias{add_grouping_column}
\title{Add grouping column to a dataset}
\usage{
add_grouping_column(data, group_cols)
}
\arguments{
\item{data}{A data.table, data.frame or tibble object containing the data to
which to add the grouping column.}

\item{group_cols}{A character string or vector of character strings
indicating the column(s) in the data to combined into a single grouping.}
}
\value{
A tibble containing an additional grouping column (called 'grouping')
  which for every row contains a summary of the values in each of the
  specified group columns.
}
\description{
\code{add_grouping_column} is a function to combine one or more grouping
columns into a single grouping column.
}
\details{
The function adds a combined grouping column to a dataset, based on a set of
specified group columns, which shows for every row which value each group
column has.
}
\examples{
add_grouping_column(
  data = dummy_gasprice,
  group_cols = c("state", "oil_company")
)
}
