% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIERARCHICAL_add_hierarchical_lineage.R
\name{add_hierarchical_lineage}
\alias{add_hierarchical_lineage}
\title{Add hierchical lineage to the data}
\usage{
add_hierarchical_lineage(data, hierarchical_col = "")
}
\arguments{
\item{data}{A tibble containing information on the hierarchy structure. There
should be two variables, one giving the name of the group and the other
it's level. The order can be descending (top down) or descending (bottom
up), but the resulting output will be descending (top down) for
readability.}

\item{hierchical_col}{A character string indicating which column corresponds
to the hierarchy group for which the lineage should be added to the data.}
}
\value{
A tibble extended with the lineage for the specified hierchical
  group.
}
\description{
\code{add_hierarchical_lineage} is a function that identifies both the direct
parents of every node, as well as the ultimate leaf groups in a hierarchy.
}
\examples{
dummy_hierarchical_gasprice \%>\%
   dplyr::select(location, level_location) \%>\%
   dplyr::distinct() \%>\%
   add_hierarchical_lineage(hierarchical_col = "location")
}
