% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agemodel.R
\name{agemodel}
\alias{agemodel}
\title{Draw a random, strictly increasing age model given a vector 'ages' and
their associated 1 sigma uncertainties ('sigmas') according to a
Gaussian. Requires that the difference between time steps should be
within a ratio firstdiffagreementratio' (default to 1, or 100%).}
\usage{
agemodel(ages, sigmas, n.sigma = 2, firstdiffagreementratio = 1,
  tolerance = 10^-(12))
}
\arguments{
\item{ages}{A vector containing the age data.}

\item{sigmas}{A vector containing the 1 sigma uncertainties associated with the age data.}

\item{firstdiffagreementratio}{How large can the slope between time steps be? Defaults to 1, or 100%.}

\item{tolerance}{A tolerance level to speed up computations when age points are very close.}
}
\description{
Draw a random, strictly increasing age model given a vector 'ages' and
their associated 1 sigma uncertainties ('sigmas') according to a
Gaussian. Requires that the difference between time steps should be
within a ratio firstdiffagreementratio' (default to 1, or 100%).
}
