% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agemodel.R
\name{agemodels}
\alias{agemodels}
\title{A wrapper around agemodel() that returns a matrix of multiple age models.
Each column is an age model.}
\usage{
agemodels(ages, sigmas, n.models = 2, n.sigma = 2,
  firstdiffagreementratio = 1, tolerance = 10^-(10))
}
\arguments{
\item{ages}{A vector containing the age data.}

\item{sigmas}{A vector containing the 1 sigma uncertainties associated with the age data.}

\item{firstdiffagreementratio}{How large can the slope between time steps be? Defaults to 1, or 100%.}

\item{tolerance}{A tolerance level to speed up computations when age points are very close.}
}
\description{
A wrapper around agemodel() that returns a matrix of multiple age models.
Each column is an age model.
}
