% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctFunctions.R
\name{annpct}
\alias{annpct}
\title{Percentage function}
\usage{
annpct(x)
}
\arguments{
\item{x}{object of class time series}
}
\description{
Computes the annual percentage change at each time point.
}
\examples{
ts_1 <- ts(rnorm(12), frequency=4, start=c(2004,1))
annpct(ts_1)
}
\author{
Florian Eckert and Heiner Mikosch, KOF, ETH Zurich
}

