% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_forecast.R
\name{auto_dif}
\alias{auto_dif}
\title{Test and treat non-stationary data}
\usage{
auto_dif(data, type = c("kpss", "adf", "pp"), alpha = 0.05, exclude = NULL)
}
\arguments{
\item{data}{A tbl with variables in columns.}

\item{type}{Vector with tests to perform. Available: list("kpss", "pp, "adf")}

\item{alpha}{Significance level for the tests.}

\item{exclude}{Variables to exclude from evaluation.}

}

\value{
An object of class list with a tibble (series_dif) containing the non-stationary data in first differences; and a tibble (series_ndiffs) containing the results from the unit roots test performed.
}
\description{
This function tests for unit roots and applies first differences to treat non-stationary data.
}
\details{
Tests available are c("kpss", "adf", "pp").
}
\note{
Only first differences are taken, even if series are integrated of order > 1. So it is advisable to look at "series_ndiffs" to assess the results from the unit root tests and, if necessary, run the function again until all the series are I(0).
}
\examples{
\dontrun{
library(tidyverse)
library(tstools)
mu <- list(-2, 0, 2, 4, 6)
sigma <- list(1, 2, 3, 4, 5)
series <- purrr::map2_dfc(mu, sigma, rnorm, n = 100)
series <- dplyr::mutate_at(series, vars(V1,V3), funs(cumsum))
series_dif <- auto_dif(series, type = c("kpss", "pp"))
}
}
\author{
J. Renato Leripio
}
