% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{bam2coverage}
\alias{bam2coverage}
\title{Convert a bam file to coverage}
\usage{
bam2coverage(file.id, type = c("SINGLE", "PAIRED"), fragment.length = 200,
  chr.flt = NA)
}
\arguments{
\item{file.id}{path of the bam file}

\item{type}{sequencing library type "SINGLE" or "PAIRED"}

\item{fragment.length}{average chromatin fragment size, default 200}

\item{chr.flt}{vector of chromosome names to be included}
}
\value{
a coverage object (\code{\link[IRanges]{RleList}} as returned by \code{\link[IRanges]{coverage}})
}
\description{
Reads can optionally be filtered by a chromosome inclusion list
Single reads will be extended to the provided fragment length
}

