% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomes.R
\name{bed2dyad}
\alias{bed2dyad}
\title{Convert a bed file to nucleosome dyad coverage}
\usage{
bed2dyad(file.id, type = c("SINGLE", "PAIRED"), width = 1)
}
\arguments{
\item{file.id}{path of the bed file}

\item{type}{sequencing library type "SINGLE" or "PAIRED"}

\item{width}{width of the dyad, default 1}
}
\value{
a coverage object (\code{\link[IRanges]{RleList}} as returned by \code{\link[IRanges]{coverage}})
}
\description{
Single reads will be shifted towards the dyad position assuming an end-localisation of the read
For paired reads the fragment mid points will be considered dyad position
}
\details{
the bed files for SINGLE should have columns 1) chromosome 2) start 3) end 4) strand
the bed files for PAIRED should have columns 1) chromosome 2) start 3) end
}

