% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomes.R
\name{bin.matrix}
\alias{bin.matrix}
\title{Bin a matrix column-wise and average the bins across rows}
\usage{
bin.matrix(m, bin.size)
}
\arguments{
\item{m}{a matrix}

\item{bin.size}{bin size}
}
\value{
a vector of binned values
}
\description{
Bin a matrix column-wise and average the bins across rows
}

