% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomes.R
\name{bin.matrix.rows}
\alias{bin.matrix.rows}
\title{Bin a matrix column-wise per row}
\usage{
bin.matrix.rows(m, bin.size)
}
\arguments{
\item{m}{a matrix}

\item{bin.size}{bin size}
}
\value{
a matrix of binned rowws
}
\description{
Bin a matrix column-wise per row
}

