% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Bin a dataset by a common column.}
\usage{
bin(dt, bin.size, bin.column, bin.average.function = mean.narm,
  interpolate = T, remove.na = T, time.sampled.at = "mid")
}
\arguments{
\item{dt}{A data frame containing the data to be binned.}

\item{bin.size}{The size of the bins}

\item{bin.column}{The column to use for binning.}

\item{bin.average.function}{The function to use for averaging bins. Default is the arithmetic mean.}

\item{interpolate}{Should empty bins be interpolated linearly?}

\item{remove.na}{Should NAs remaining after interpolation (usually at endpoints after interpolatin) be removed? BEWARE: be careful about removing nans before interpolating.}

\item{time.sampled.at}{How time indices are constructed. Either 'start', 'mid' or 'end' of bins. Defaults to 'mid'.}
}
\description{
Bin a dataset by a common column.
}
