% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_equallyspaced.R
\name{bin_equallyspaced_df}
\alias{bin_equallyspaced_df}
\title{Create equally spaced bins for a column of a data frame.}
\usage{
bin_equallyspaced_df(df, by = NULL, bin.size, include.lowest = T)
}
\arguments{
\item{df}{The data frame.}

\item{by}{The column of the data frame.}

\item{bin.size}{The size of the bins.}

\item{include.lowest}{Include endpoint bin?}
}
\description{
Create equally spaced bins for a column of a data frame.
}
