% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_equallyspaced.R
\name{bin_equallyspaced_vector}
\alias{bin_equallyspaced_vector}
\title{Create equally spaced bins for a vector.}
\usage{
bin_equallyspaced_vector(v, bin.size, include.lowest = T)
}
\arguments{
\item{v}{The vector.}

\item{bin.size}{The size of the bins.}
}
\description{
Create equally spaced bins for a vector.
}
