% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotmaximum.R
\name{boxplot.maximum}
\alias{boxplot.maximum}
\title{Find the largest v less than or equal to
upper hinge of boxplot, which is
Q1 (25th quantile) + 1.5 * IQR}
\usage{
\method{boxplot}{maximum}(v)
}
\arguments{
\item{v}{A scalar valued vector.}
}
\value{
The maximum value for a boxplot based on 'v'.
}
\description{
Find the largest v less than or equal to
upper hinge of boxplot, which is
Q1 (25th quantile) + 1.5 * IQR
}
