% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot-minimum.R
\name{boxplot_minimum}
\alias{boxplot_minimum}
\title{Find the smallest v greater than or equal to
lower hinge of boxplot, which is
Q1 (25th quantile) - 1.5 * IQR}
\usage{
boxplot_minimum(v)
}
\arguments{
\item{v}{A scalar valued vector.}
}
\value{
The minimum value for a boxplot based on 'v'.
}
\description{
Find the smallest v greater than or equal to
lower hinge of boxplot, which is
Q1 (25th quantile) - 1.5 * IQR
}
