% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantons.R
\name{cantons}
\alias{cantons}
\title{Vector of Swiss cantons}
\usage{
cantons(full = F)
}
\arguments{
\item{full}{logical full names or two letter classification. defaults to FALSE.}
}
\description{
Vector of Swiss cantons
}
