% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm.R
\name{ccm}
\alias{ccm}
\title{Performs CCM for a given lag.}
\usage{
ccm(data, library.sizes = 100, convergence.test = T, lag = 0, E = 2,
  tau = 1, lib = c(1, dim(data)[1]), pred = lib, samples.original = 100,
  samples.surrogates = 50, num.neighbours = E + 1, random.libs = TRUE,
  with.replacement = TRUE, exclusion.radius = E + 1, epsilon = NULL,
  silent = TRUE, RNGseed = 1111, surrogate.method = "AAFT",
  n.surrogates = 0, parallel = FALSE, library.column = 1,
  target.column = 2, surrogate.column = 2, print.to.console = T,
  print.surrogate.to.console = T, n.libsizes.to.check = 20)
}
\description{
Performs CCM for a given lag.
}
