% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm-lagged.R
\name{ccm_lagged}
\alias{ccm_lagged}
\title{Cross-mapping from one variable to another variable over a range of lags.
Wrapper around OnewayLaggedCCM allowing multiple values for each input argument.}
\usage{
ccm_lagged(data, lags, Es = NULL, taus = NULL,
  library.sizes = c(as.integer(nrow(data)/2)), lib = c(1, dim(data)[1]),
  pred = lib, samples = 100, samples.surrogates = 50, n.surrogates = 0,
  surrogate.methods = c("aaft"), time.unit = NULL, time.bin.size = NULL,
  num.neighbours = E + 1, random.libs = TRUE, with.replacement = TRUE,
  exclusion.radius = E + 1, epsilon = NULL, RNGseed = 1111,
  silent = TRUE, time.run = F, print.to.console = T,
  time.series.length.threshold = 100, library.column = 1,
  target.column = 2, surrogate.column = target.column,
  convergence.test = TRUE, parallel = TRUE, parallelize.on.each.lag = F,
  num.cores = parallel::detectCores() - 1, regression.convergence.plots = F,
  always.run.surrogates = F, n.libsizes.to.check = 20,
  optimise.FNNdim = F, optimise.boxcountdim = F, min.E = 2, max.E = 10,
  min.tau = 1, max.tau = 1, plot.simplex.projection = F, ...)
}
\arguments{
\item{data}{A data frame containing two columns - one for the presumed driver and one for the response.}

\item{lags}{A vector of lags to compute CCM for.}

\item{library.sizes}{The size of the random libraries drawn when calculating cross map skill.}

\item{lib}{Indices of the original library time series to use as the library (training) set.}

\item{pred}{Indices of the original target time series to use as prediction set. If this overlaps
with the training set, make sure to use leave-K-out cross validation setting the
'exclusion.radius' parameters to a minimum of E + 1.}

\item{samples}{The number of random libraries to draw when calculating cross map skill.}

\item{samples.surrogates}{The number of surrogate series in each null ensemble.}

\item{n.surrogates}{Should a surrogate test also be performed? If so, 'n.surrogates' sets
 the number of surrogate time series to use. By default, no surrogate test is performed
(n.surrogates = 0).}

\item{time.unit}{The time unit of the raw time series.}

\item{time.bin.size}{The temporal resolution of the raw time series (given in the units
indicated by 'time.unit').}

\item{num.neighbours}{The number of nearest neighbours to use in predictions. Defaults to E + 1.}

\item{random.libs}{Whether or not to sample random library (training) sets. Defaults to TRUE.}

\item{with.replacement}{Should samples be drawn with replacement? Defaults to TRUE.}

\item{exclusion.radius}{The number of temporal neighbours to exclude for the leave-K-out cross
validation.}

\item{RNGseed}{For reproducivility. Seed to use for the random number generator.}

\item{time.series.length.threshold}{Display a warning if the time series length drops
below this threshold.}

\item{library.column}{Integer indicating which column to use as the library column (presumed response)
(1 for the first column and 2 for the second column).}

\item{target.column}{Integer indicating which column to use as the target column (presumed driver).
Defaults to the opposite of 'library.column'.}

\item{surrogate.column}{Which column to use to generate surrogates. Defaults to the value of
'target.column' (the presumed driver).}

\item{convergence.test}{Should a convergence test be performed? Analyses where CCM does not
convergence are nonsensical, so this option defaults to TRUE.}

\item{parallelize.on.each.lag}{Should parallellisation be done on the outer lag loop?
Defaults to TRUE. Otherwise, parallellisation is done over the surrogate analyses.}

\item{num.cores}{The number of CPU cores to use for parallelisation. Defaults to one
core less than what is available.}

\item{regression.convergence.plots}{Display regression plots for the convergence test?
Defaults to FALSE.}

\item{always.run.surrogates}{Should surrogate analyses be performed even if the convergence test
fails? Defaults to FALSE (there is no reason to perform significance testing if the analysis is not
causal to begin with).}

\item{n.libsizes.to.check}{How many library sizes to use for the convergence check?
Defaults to 30.}

\item{optimise.FNNdim}{Optimise false nearest neighbours? Defaults to FALSE, meaning that simplex projection only is used
for optimisation. If TRUE, simplex projection is performed with the minimum embedding dimension outputted from
the FNN algorithm (and the box counting dimension, if acitivated). Only relevant if either 'E' or tau is set to NULL.}

\item{optimise.boxcountdim}{Optimise box counting dimension? Defaults to FALSE, meaning that simplex projection only is used
for optimisation. If TRUE, simplex projection is performed with the minimum embedding dimension as the first integer
larger than twice the box counting algorithm of the attractor (also considering the minimum dimension indicated by the
FNN algorithm, if activated). Only relevant if either 'E' or tau is set to NULL.}

\item{min.E}{The minimum embedding dimension for which to optimise 'E'. Only relevant if either 'E' or tau is set to NULL.}

\item{max.E}{The maximum embedding dimension for which to optimise 'E'. Only relevant if either 'E' or tau is set to NULL.}

\item{min.tau}{The minimum embedding lag for which to optimise 'tau'. Only relevant if either 'E' or tau is set to NULL.}

\item{max.tau}{The maximum embedding lag for which to optimise 'tau'. Only relevant if either 'E' or tau is set to NULL.}

\item{E}{The embedding dimension. Defaults to NULL, which triggers automated optimisation of the embedding
dimension up to the dimension specified by 'max.E'.}

\item{tau}{The embedding lag. Defaults to NULL, which triggers automated optimisation of the embedding lag
up to the dimension specified by 'max.tau'. For sparsely sampled time series (for example geological
time series), it is wise to set this value to 1. For densely sampled time series, this should be set to
the first minima of the autocorrelation function of the presumed driver.}

\item{surrogate.method}{The type of surrogate time series to generate. Will vary depending
on what null hypothesis is being tested. Defaults to AAFT surrogates.}

\item{parallell}{Activate parallellisation? Defaults to true. Currently, this only works
decently on Mac and Linux systems.}
}
\description{
Cross-mapping from one variable to another variable over a range of lags.
Wrapper around OnewayLaggedCCM allowing multiple values for each input argument.
}
