% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm-lagged.R
\name{ccm_lagged_oneway}
\alias{ccm_lagged_oneway}
\title{Cross-mapping from one variable to another variable over a range of lags.}
\usage{
ccm_lagged_oneway(data, lags, E = NULL, tau = NULL,
  library.sizes = as.integer(nrow(data)/2), lib = c(1, dim(data)[1]),
  pred = lib, samples = 100, samples.surrogates = 50, n.surrogates = 0,
  surrogate.method = "AAFT", always.run.surrogates = F, time.unit = NULL,
  time.bin.size = NULL, num.neighbours = E + 1, random.libs = TRUE,
  with.replacement = TRUE, exclusion.radius = ifelse(is.null(E), yes = 10,
  no = E + 1), epsilon = NULL, RNGseed = 1111, silent = TRUE,
  time.run = F, print.to.console = T, time.series.length.threshold = 100,
  library.column = 1, target.column = 2, surrogate.column = target.column,
  convergence.test = TRUE, parallel = TRUE, parallelize.on.each.lag = F,
  num.cores = parallel::detectCores() - 1, n.libsizes.to.check = 30,
  regression.convergence.plots = F, max.E = 10, max.tau = 1)
}
\arguments{
\item{data}{A data frame containing two columns - one for the presumed driver and one for the response.}

\item{lags}{A vector of lags to compute CCM for.}

\item{E}{The embedding dimension. Defaults to NULL, which triggers automated optimisation of the embedding
dimension up to the dimension specified by 'max.E'.}

\item{tau}{The embedding lag. Defaults to NULL, which triggers automated optimisation of the embedding lag
up to the dimension specified by 'max.tau'. For sparsely sampled time series (for example geological
time series), it is wise to set this value to 1. For densely sampled time series, this should be set to
the first minima of the autocorrelation function of the presumed driver.}

\item{library.sizes}{The size of the random libraries drawn when calculating cross map skill.}

\item{lib}{Indices of the original library time series to use as the library (training) set.}

\item{pred}{Indices of the original target time series to use as prediction set. If this overlaps
with the training set, make sure to use leave-K-out cross validation setting the
'exclusion.radius' parameters to a minimum of E + 1.}

\item{samples}{The number of random libraries to draw when calculating cross map skill.}

\item{samples.surrogates}{The number of surrogate time series in the null
ensemble.}

\item{n.surrogates}{Should a surrogate test also be performed? If so, 'n.surrogates' sets
 the number of surrogate time series to use. By default, no surrogate test is performed
(n.surrogates = 0).}

\item{surrogate.method}{The type of surrogate time series to generate. Will vary depending
on what null hypothesis is being tested. Defaults to AAFT surrogates.}

\item{always.run.surrogates}{Should surrogate analyses be performed even if the convergence test
fails? Defaults to FALSE (there is no reason to perform significance testing if the analysis is not
causal to begin with).}

\item{time.unit}{The time unit of the raw time series.}

\item{time.bin.size}{The temporal resolution of the raw time series (given in the units
indicated by 'time.unit').}

\item{num.neighbours}{The number of nearest neighbours to use in predictions. Defaults to E + 1.}

\item{random.libs}{Whether or not to sample random library (training) sets. Defaults to TRUE.}

\item{with.replacement}{Should samples be drawn with replacement? Defaults to TRUE.}

\item{exclusion.radius}{The number of temporal neighbours to exclude for the leave-K-out cross
validation.}

\item{epsilon}{Exlude neighbours if the are within a distance of 'epsilon'
from the predictee.}

\item{RNGseed}{For reproducivility. Seed to use for the random number generator.}

\item{silent}{Suppress warnings?}

\item{time.run}{Time the run?}

\item{print.to.console}{Display progress?}

\item{time.series.length.threshold}{Display a warning if the time series length drops
below this threshold.}

\item{library.column}{Integer indicating which column to use as the library column (presumed response)
(1 for the first column and 2 for the second column).}

\item{target.column}{Integer indicating which column to use as the target column (presumed driver).
Defaults to the opposite of 'library.column'.}

\item{surrogate.column}{Which column to use to generate surrogates. Defaults to the value of
'target.column' (the presumed driver).}

\item{convergence.test}{Should a convergence test be performed? Analyses where CCM does not
convergence are nonsensical, so this option defaults to TRUE.}

\item{parallel}{Activate parallellisation? Defaults to true. Currently, this only works
decently on Mac and Linux systems.}

\item{parallelize.on.each.lag}{Should parallellisation be done on the outer lag loop?
Defaults to TRUE. Otherwise, parallellisation is done over the surrogate analyses.}

\item{num.cores}{The number of CPU cores to use for parallelisation. Defaults to one
core less than what is available.}

\item{n.libsizes.to.check}{How many library sizes to use for the convergence check?
Defaults to 30.}

\item{regression.convergence.plots}{Display regression plots for the convergence test?
Defaults to FALSE.}

\item{max.E}{The maximum embedding dimension for which to optimise 'E'.}

\item{max.tau}{The maximum embedding lag for which to optimise 'tau'.}
}
\description{
Cross-mapping from one variable to another variable over a range of lags.
}
