% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm-over-library-sizes.R
\name{ccm_over_library_sizes}
\alias{ccm_over_library_sizes}
\title{importFrom magrittr "%>%"}
\usage{
ccm_over_library_sizes(lag, data, E = 2, tau = 1, library.sizes = 100,
  low.libsize = min(E * tau + max(2, abs(lag)), 10, 20, na.rm = T),
  n.libsizes.to.check = 30, high.libsize = min(ceiling(max(library.sizes) *
  1.5), max(library.sizes) - E * tau - abs(lag)), lib = c(1, dim(data)[1]),
  pred = lib, samples.original = 300, samples.surrogates = 0,
  n.surrogates = 0, surrogate.method = "AAFT", time.unit = "bins",
  time.bin.size = 1, num.neighbours = E + 1, random.libs = TRUE,
  with.replacement = TRUE, exclusion.radius = E, epsilon = NULL,
  RNGseed = 1111, silent = TRUE, parallel = F, time.run = F,
  print.to.console = F, time.series.length.threshold = 100,
  library.column = 1, target.column = 1, surrogate.column = target.column)
}
\arguments{
\item{library.sizes}{Either a single maximum library size (cross mapping
is performed for a range of value from the smallest possible library
size to the provided library size) or a user-specified range
of library sizes. If user-provided, make sure to provide at least
20 different library sizes to ensure robust convergence assessment.}

\item{low.libsize}{If one library size is specified, cross map
for library sizes ranging from 'low.libsize' to 'high.libsize'.}

\item{high.libize}{If one library size is specified, cross map
for library sizes ranging from 'low.libsize' to 'high.libsize'.}
}
\description{
importFrom magrittr "%>%"
}
