% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-circulant_embedding.R
\name{ce_surrogate}
\alias{ce_surrogate}
\title{Generate a surrogate series using Davies and Harte's circulant embedding.
The values of the surrogate series are not constrained to the values of
the original time series.}
\usage{
ce_surrogate(series)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}
}
\description{
Generate a surrogate series using Davies and Harte's circulant embedding.
The values of the surrogate series are not constrained to the values of
the original time series.
}
