% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcFunctions.R
\name{chainedGrc}
\alias{chainedGrc}
\title{Weight real growth rates by their respective nominal weights}
\usage{
chainedGrc(rlist, nlist)
}
\arguments{
\item{rlist}{list of time series object}

\item{nlist}{list of time series object}
}
\description{
Computes growth contributions by weighting real growth rates by average nominal gdp of the previous year
}
\examples{
mat_r <- matrix(c(7+rnorm(12),50+rnorm(12),70+rnorm(12),100+rnorm(12)),ncol=4)
r_ts <- ts(mat_r, frequency=4, start=c(2004,1))
colnames(r_ts) <- c("IT","DE","CH","FR")
rlist<-as.list(r_ts)
mat_n <- matrix(c(7+rnorm(12),50+rnorm(12),70+rnorm(12),100+rnorm(12)),ncol=4)
n_ts <- ts(mat_n, frequency=4, start=c(2004,1))
colnames(n_ts) <- c("IT","DE","CH","FR")
nlist<-as.list(n_ts)
chainedGrc(rlist,nlist)
}
\author{
Florian Eckert and Heiner Mikosch, KOF, ETH Zurich
}

