% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged-ccm-convergence.R
\name{check_ccm_convergence}
\alias{check_ccm_convergence}
\title{importFrom magrittr "%>%"}
\usage{
check_ccm_convergence(lag, data, E = 2, tau = 1, low.libsize = min(E * tau
  + max(2, abs(lag)), 10, 20, na.rm = T), library.size = 100,
  n.libsizes.to.check = 30, high.libsize = min(ceiling(library.size * 1.5),
  library.size - E * tau - abs(lag)), lib = c(1, dim(data)[1]), pred = lib,
  samples.original = 1000, samples.surrogates = 100, n.surrogates = 0,
  surrogate.method = "AAFT", time.unit = "bins", time.bin.size = 1,
  num.neighbours = E + 1, random.libs = TRUE, with.replacement = TRUE,
  exclusion.radius = E, epsilon = NULL, RNGseed = 1111, silent = TRUE,
  parallel = F, time.run = F, print.to.console = F,
  time.series.length.threshold = 100, library.column = 1,
  target.column = 1, surrogate.column = target.column,
  plot.regressions = T)
}
\description{
importFrom magrittr "%>%"
}
