% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MISC_check_data_format.R
\name{check_data_format}
\alias{check_data_format}
\title{Checks specified data to have a required format.}
\usage{
check_data_format(
  data,
  func_name = "this function",
  req_cols = c(),
  unique_value_cols = c()
)
}
\arguments{
\item{data}{A tibble containing the data to be checked in terms of required
format.}

\item{req_cols}{A vector of character strings corresponding to the column(s)
for which it is checked that they are available in the data.}

\item{unique_value_cols}{A vector of character strings corresponding to the
column(s) for which it is checked that they contain only a single unique
value.}
}
\value{
Nothing.
}
\description{
\code{check_data_format} is a function to check whether the specified data
has the same format as is expected by the function in which this check is
performed.
}
\examples{
check_data_format(dummy_gasprice)
}
