% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHGDP.R
\docType{data}
\name{CHGDP}
\alias{CHGDP}
\title{CH GDP Growth Contributions}
\format{
List list of six time series of class ts,
containing contributions to Swiss GDP growth
\describe{
  \item{manufacturing}{Growth contribution of manufacturing.}
  \item{energy}{Growth contribution of energy, water sector}
  \item{construction}{Growth contribution construction sector.}
  \item{hotels}{Growth contribution of hotels.}
  \item{fin_insur}{Growth contribution of financial services and insurances.}
  \item{other}{Growth contribution of other sectors.}
}
}
\source{
\url{https://www.seco.admin.ch/seco/en/home/wirtschaftslage---wirtschaftspolitik/Wirtschaftslage/bip-quartalsschaetzungen-/daten.html}
}
\usage{
CHGDP
}
\description{
A list of time series containing sector contributions to Swiss GDP over time.
}
\keyword{datasets}
