% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_ts.R
\name{concat_ts}
\alias{concat_ts}
\title{Concatenate to Non-Overlapping Time Series}
\usage{
concat_ts(ts1, ts2)
}
\arguments{
\item{ts1}{object of class ts1, typically the older of two time series.}

\item{ts2}{object of class ts1, typically the younger of two time series.}
}
\description{
Append one time series to another. This only works for non-overlapping time series of the same frequency.
For overlapping time series please see \code{\link{resolveOverlap}}.
}
