% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convertToColors}
\alias{convertToColors}
\title{Convert numeric matrix to colors}
\usage{
convertToColors(mat, rng)
}
\arguments{
\item{mat}{a matrix of values}

\item{rng}{mapping range}
}
\value{
a matrix of colors
}
\description{
Useful for plotting heatmaps
}

