% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{coverageWindowsCenteredStranded}
\alias{coverageWindowsCenteredStranded}
\title{Get windowed strand-oriented coverages around center points}
\usage{
coverageWindowsCenteredStranded(centers, window.size = 1000, coverage)
}
\arguments{
\item{centers}{a data frame of center points with columns 'chr', 'center', 'strand'}

\item{window.size}{the size of the window surrounding the center position, default 1000}

\item{coverage}{a coverage object (\code{\link[IRanges]{RleList}} as returned by \code{\link[IRanges]{coverage}})}
}
\value{
a matrix
}
\description{
Out-of-bound windows will be removed!
}
