% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{coverageWindowsStranded}
\alias{coverageWindowsStranded}
\title{Get windowed strand-oriented coverages}
\usage{
coverageWindowsStranded(windows, coverage)
}
\arguments{
\item{coverage}{a coverage object (\code{\link[IRanges]{RleList}} as returned by \code{\link[IRanges]{coverage}})}

\item{windwos}{a data frame of windows with columns 'chr', 'start', 'end', 'strand'}
}
\value{
a matrix
}
\description{
Get windowed strand-oriented coverages
}
\examples{
coverageWindowsCenteredStranded()

}

