% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIERARCHICAL_create_hierarchy_list.R
\name{create_hierarchy_list}
\alias{create_hierarchy_list}
\title{Create hierarchy list}
\usage{
create_hierarchy_list(data)
}
\arguments{
\item{data}{The initialized forecast data with the hierarchical "level_"
column in it, which has been created using the
\code{initialize_ts_forecast_data} function.}
}
\value{
A named list of two objects - the hierarchy matrix ("matrix") and
  data with the hierarhical group names and their respective levels (
  "data").
}
\description{
\code{create_hierarchy_list} is a function that creates the named list
"hierarchy", which includes both the hierarchical matrix as well as the data
used to create that matrix.
}
\examples{
dummy_hierarchical_gasprice \%>\%
   initialize_ts_forecast_data(
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = "currency",
      hierarchical_cols = c("location", "oil_company")
   ) \%>\%
   create_hierarchy_list()
}
