% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIERARCHICAL_create_hierarchy_matrix.R
\name{create_hierarchy_matrix}
\alias{create_hierarchy_matrix}
\title{Create the hierarchy matrix}
\usage{
create_hierarchy_matrix(data)
}
\arguments{
\item{data}{A tibble containing information on the hierarchy structure. There
should be three variables: the hierarchy groups, their levels ("level") and
whether they are a leaf or not ("leaf"). The assumption is that this input
is the output of the \code{identify_hierarchy_leaves} function.}
}
\value{
A matrix of 1's and 0's where the rows contain all of the group names
  and the columns contain all of the leaves.
}
\description{
\code{create_hierarchy_matrix} is a function that creates the hierarchy
matrix. This is a matrix that stores the relevant information about the
relationship (i.e. connections) between the different groups in your
hierarchy.
}
\examples{
dummy_hierarchical_gasprice \%>\%
   dplyr::select(location, oil_company, level_location, level_oil_company) \%>\%
   dplyr::distinct() \%>\%
   add_hierarchical_lineage(hierarchical_col = "location") \%>\%
   add_hierarchical_lineage(hierarchical_col = "oil_company") \%>\%
   create_hierarchy_matrix()
}
