% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm-surrogate.R
\name{create_surrogate_dataframes}
\alias{create_surrogate_dataframes}
\title{Create a dataframe of multiple surrogate series.}
\usage{
create_surrogate_dataframes(original.data, surrogate.column,
  surrogate.method = "aaft", n.surrogates = 100, print.to.console = F)
}
\arguments{
\item{Original}{data (a two-column data frame)}

\item{Surrogate}{data (a n.observations-by-n.surrogates matrix)}

\item{surrogate_column}{The index or name of of the column for which
       to generate surrogate data. When doing causal analyses,
       this should correspond to the target column - the putative driver.
Replaces a column of a data frame with surrogate}
}
\description{
Create a dataframe of multiple surrogate series.
}
