% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{cumPlot}
\alias{cumPlot}
\title{Cumulative plot with interquartile ranges}
\usage{
cumPlot(mat, base.col = "#000000", ...)
}
\arguments{
\item{mat}{matrix or data.frame to cumulate. colum names specify x}

\item{base.col}{base color}
}
\description{
Cumulative plot with interquartile ranges
}
