% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-random-series.R
\name{datamodel_sigmas}
\alias{datamodel_sigmas}
\title{Draws random data according to a Gaussian distribution centered
around the original data with standard deviations given by
n.sigma * sigmas.}
\usage{
datamodel_sigmas(data, sigmas, n.sigma)
}
\arguments{
\item{data}{A vector of data.}

\item{sigmas}{A vector of standard deviations.}

\item{n.sigma}{The number of standard deviations for the Gaussian.}
}
\description{
Draws random data according to a Gaussian distribution centered
around the original data with standard deviations given by
n.sigma * sigmas.
}
