% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{datamodels}
\alias{datamodels}
\title{A wrapper around datamodel()}
\usage{
datamodels(data, sigmas = NA, n.sigma = 2, lower_bounds = NA,
  upper_bounds = NA, n.models = 1)
}
\arguments{
\item{data}{A vector of data.}

\item{sigmas}{A vector of 1 sigma uncertainties associated with the data.}

\item{n.sigma}{How many sigmas should be allowed?}

\item{lower_bounds}{A vector of lower bounds for the data.}

\item{upper_bounds}{A vector of upper bounds for the data.}
}
\value{
A matrix of data models
}
\description{
A wrapper around datamodel()
}
