% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_date_to_period.R
\name{date_to_period}
\alias{date_to_period}
\title{Convert date object to period}
\usage{
date_to_period(date)
}
\arguments{
\item{date}{A date object}
}
\value{
An object in yyyymm numeric format
}
\description{
\code{date_to_period} is a function to transform a date object into a period
(numeric - yyyymm format).
}
\examples{
date_to_period(as.Date("2018-05-09"))
}
\seealso{
Other Date: 
\code{\link{fast_read_RData_from_network}()},
\code{\link{get_date_range}()},
\code{\link{months_between_periods}()},
\code{\link{parse_date_from_vector}()},
\code{\link{period_to_first_day}()},
\code{\link{period_to_quarter}()}
}
\concept{Date}
