% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{default_tsplot_theme}
\alias{default_tsplot_theme}
\title{Initiate default theme}
\usage{
default_tsplot_theme()
}
\description{
The \code{\link{tsplot}} methods provide a theme argument which is used to pass on a plethora of useful defaults. These defaults are essentially stored in a list. Sometimes the user may want to tweak some of these defaults while keeping most of them.
Hence the initDefaultTheme function create a fresh list object containing default values for lot of different layout parameters etc. By replacing single elements of the list and passing the entire list to the plot function, single aspects can be tweaked while keeping most defaults. Init defaultTheme does not need any parameters.
}
\details{
Themes are essentially list that contain \code{\link{par}} parameters. Below all items are listed, some of them with comments. I will try to write comments on all params soon.
The list contains the following elements:
\itemize{
\item margins
\item fillYearWithNAs logical: should a year be filled up with NAs if values do not go until the last period of the last year?
\item line_colors vector of RGB code strings defaults
\item line_to_middle
\item lwd integer vector of line width
\item lty integer vector of line types, solid, dashed etc. see also \code{\link{par}}.
\item xaxs
\item yaxs
\item bar_border RGB color code string of all bars borders
\item total_bar_margin_pct share of space left for margins when grouping bar charts
\item bar_fill_color vector RGB color code strings denoting the fill of (stacked) bar chart elements
\item sum_as_line
\item sum_line_color
\item sum_line_lty
\item sum_line_lwd
\item highlight_window logical should a highlight window be used for background?
\item highlight_window_start integer vector year,period
\item highlight_window_end integer vector year,period
\item highlight_color RGB string color, color of the highlight background window
\item use_box logical: should there be a box around the plot? defaults to FALSE.
\item y_las
\item lwd_ticks_1
\item lwd_ticks_2
\item yearly_ticks logical: should there be yearly tick marks?
\item quarterly_ticks logical: should there be quarterly tick marks?
\item monthly_ticks logical: should there be monthly tick marks?
\item tcl_quarterly_tick_tcl
\item tcl_yearly_tick
\item lwd_yearly_ticks
\item lwd_quarterly_ticks
\item label_pos
\item show_left_y_axis
\item show_right_y_axis
\item y_grid_count
\item show_y_grids
\item y_grid_color
\item legend_col
\item title_outer
\item title_adj
\item title_line
\item title_cex.main
\item subtitle_adj
\item subtitle_outer
\item subtitle_line
\item subtitle_cex.main
\item subtitle_transform
\item subtitle_adj_r
}
}
\examples{
# create a list
data(KOF)
tt <- default_tsplot_theme()
# adjust a single element
tt$highlight_window <- TRUE
# pass the list to tsplot
tsplot(KOF$kofbarometer,theme = tt)
# for more theme examples check the vignette
vignette("tstools")


}
\author{
Matthias Bannert
}
