% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-phase-and-amplitude-randomisation.R
\name{dh_surrogate}
\alias{dh_surrogate}
\title{Generate a Davison and Hinkley's surrogate series. Randomises both
the phases and amplitudes in the frequency domain before inverting back
to the time domain. The resulting surrogates are not constrained to the
values of the original time series.}
\usage{
dh_surrogate(series)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}
}
\description{
Generate a Davison and Hinkley's surrogate series. Randomises both
the phases and amplitudes in the frequency domain before inverting back
to the time domain. The resulting surrogates are not constrained to the
values of the original time series.
}
