% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm-lagged-summarise.R
\name{directionalcausaltest}
\alias{directionalcausaltest}
\title{Summarise a lagged CCM result.}
\usage{
directionalcausaltest(res, library.size = max(res$library.size))
}
\arguments{
\item{library.size}{The library size to summarise over. Defaults to the
largest available library size.}

\item{dt}{A data frame containing the result of a tstools::ccm_lagged() call.}
}
\value{
The difference between summed negative median CCM skills and
    summed positive median CCM skills. If the sum > 0, then the analysis
    passes the lag test.
}
\description{
Summarise a lagged CCM result.
}
\details{
CCM predictability for a given lag is only considered if the analysis is
   convergent. Convergence is determined by Wilcoxon rank-sum test on the smallest and largest library sizes. If CCM
   skill for the largest library sizes is higher (p < 0.01) than for the
   lowest library sizes, then the analysis is potentially convergent.
   Additionally, the median CCM curve over increasing library sizes must be
   satisfactorily fitted by a convergent exponential function. If both these
   criteria are met, then the analysis is labelled convergent. The only grouping variable in this summary function is the lag, so all
   further groupings must be handled outside this function.
}
\examples{
# res is the result of a tstools::ccm_lagged() call
library(dplyr)
res \%>\% directionalcausaltest

# res is a data frame containing vertically concatenated results of
# tstools::ccm_lagged() calls for different causal variables.
res \%>\%
  group_by(causal.direction) \%>\%
  do(directionalcausaltest(.))
}
