% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swissdata_meta.R
\name{.read_swissdata_meta_unknown_format}
\alias{.read_swissdata_meta_unknown_format}
\title{Read Meta Data File w/o File Extension}
\usage{
.read_swissdata_meta_unknown_format(path)
}
\arguments{
\item{path}{character file path.}
}
\value{
Meta list if file could be located, empty list otherwise
}
\description{
Read a meta file without extension -> unknown format
Tries to determine format (yaml, json) and return the metadata
path must point to the file without extension e.g. swissdata_wd/set_id/set_id
}
