% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-ebisuzaki.R
\name{ebisuzaki_surrogate}
\alias{ebisuzaki_surrogate}
\title{Generate a phase-randomised surrogate according to Ebisuzaki.}
\usage{
ebisuzaki_surrogate(series)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}
}
\description{
Generate a phase-randomised surrogate according to Ebisuzaki.
}
