% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NETWORK_fast_read_RData_from_network.R
\name{fast_read_RData_from_network}
\alias{fast_read_RData_from_network}
\title{Fast read RData from network}
\usage{
fast_read_RData_from_network(file_path, verbose = FALSE)
}
\arguments{
\item{file_path}{A character string indicating the valid path to the file
from which the data is to be loaded (location + name of the file).}

\item{verbose}{A boolean indicating whether the function should print its
progress or not.}
}
\value{
A data.table, data.frame or tibble object containing the data that
  was stored in the RData file.
}
\description{
\code{fast_read_RData_from_network} is a function to more quickly read data
from network locations.
}
\details{
The function copies an RData file to local disk, before reading the file into
memory. Especially for larger files, this is (much) faster than directly
reading from a network location.
}
\examples{
# REPLACE network_location_path/file_name WITH A VALID FILE ON A NETWORK LOCATION
fast_read_RData_from_network(
  file_path = "network_location_path/file_name.RData",
  verbose = TRUE
)
}
\seealso{
Other Date: 
\code{\link{date_to_period}()},
\code{\link{get_date_range}()},
\code{\link{months_between_periods}()},
\code{\link{parse_date_from_vector}()},
\code{\link{period_to_first_day}()},
\code{\link{period_to_quarter}()}
}
\concept{Date}
