% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_year_with_nas.R
\name{fill_year_with_nas}
\alias{fill_year_with_nas}
\title{Fill Up a Time Series with NAs}
\usage{
fill_year_with_nas(x, add_periods = 1, fill_up_start = FALSE)
}
\arguments{
\item{x}{object of class ts}

\item{add_periods}{integer periods to add.}

\item{fill_up_start}{logical should start year be filled up? Defaults to FALSE.}
}
\description{
When plotting a time series you might want set the range of the plot a little wider than just the start and end date of the original series. This function add fills up the current period (typically year) with NA.
}
