% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillUpYearWithNAs.R
\name{fillUpYearWithNAs}
\alias{fillUpYearWithNAs}
\title{Fill Up a Time Series with NAs}
\usage{
fillUpYearWithNAs(x)
}
\arguments{
\item{ts}{object of class ts}
}
\description{
When plotting a time series you might want set the range of the plot a little wider than just the start and end date of the original series. This function add fills up the current period (typically year) with NA.
}

