% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{find_troughs}
\alias{find_troughs}
\title{Finds peaks in a time series by using a sliding window.
Wrapper around find_local_maxima().}
\usage{
find_troughs(x, npoints = 3)
}
\arguments{
\item{x}{The vector of numbers for which to identify peaks}

\item{npoints}{The number of points to either side of the local maxima.

Author: user 'stas g' on stackexchange at
https://stats.stackexchange.com/questions/22974/how-to-find-local-peaks-valleys-in-a-series-of-data}
}
\description{
Finds peaks in a time series by using a sliding window.
Wrapper around find_local_maxima().
}
