% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-acf-minima.R
\name{first_acf_minima}
\alias{first_acf_minima}
\title{Finds the first minima or zero of the autocorrelation function of the
series 'v'.}
\usage{
first_acf_minima(v, lag.max, plot.acf = F)
}
\arguments{
\item{v}{A numeric vector containing the series.}

\item{lag.max}{The maximum lag to consider for the autocorrelation
function.}

\item{plot.acf}{Should the autocorrelation function be plotted?}
}
\value{
The first lag that minimises the autocorrelation function.
}
\description{
Finds the first minima or zero of the autocorrelation function of the
series 'v'.
}
