% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-mi-minima.R
\name{first_mi_minima}
\alias{first_mi_minima}
\title{Finds the first minima or zero of the mutual information (mi) function
of the series 'v'.}
\usage{
first_mi_minima(v, lag.max, partitions = 16, plot.mi.func = F)
}
\arguments{
\item{v}{A numeric vector containing the series.}

\item{lag.max}{The maximum lag to consider in the mi function.}

\item{plot.mi}{Should the mi function be plotted?}
}
\value{
The first lag that minimises the mi function.
}
\description{
Finds the first minima or zero of the mutual information (mi) function
of the series 'v'.
}
