% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUMERIC_format_big_n.R
\name{format_big_n}
\alias{format_big_n}
\title{Format big numbers}
\usage{
format_big_n(x, digits = 2)
}
\arguments{
\item{x}{A numeric value to be formatted.}

\item{digits}{A positive or negative integer value indicating the number of
digits to use when formatting the values.}
}
\value{
A character string with big number format.
}
\description{
\code{format_big_n} is a function to format big numbers in a pretty way, to
be used when plotting.
}
\examples{
format_big_n(123456)
format_big_n(-123456789)
format_big_n(1234567890)
}
\concept{}
