% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-series.R
\name{fractal_package_surrogate}
\alias{fractal_package_surrogate}
\title{Simple wrapper for the fractal::surrogate function, to enable creating
multiple surrogate time series in one call.}
\usage{
fractal_package_surrogate(ts, method = "aaft")
}
\description{
Simple wrapper for the fractal::surrogate function, to enable creating
multiple surrogate time series in one call.
}
