% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate.R
\name{FractalPackageSurrogates}
\alias{FractalPackageSurrogates}
\title{Simple wrapper for the fractal::surrogate function, to enable creating
multiple surrogate time series in one call.}
\usage{
FractalPackageSurrogates(ts, method = "aaft")
}
\description{
Simple wrapper for the fractal::surrogate function, to enable creating
multiple surrogate time series in one call.
}
