% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_get_column_features.R
\name{get_column_features}
\alias{get_column_features}
\title{Get the features of the column(s) in the dataset}
\usage{
get_column_features(data, extended = TRUE)
}
\arguments{
\item{data}{A data.table, data.frame or tibble object containing the data to
which to get the column features.}

\item{extended}{Boolean. If TRUE (default), then the number of unique and NA
values for each column is also included in the output.}
}
\value{
A tibble containing the name(s) of the column(s) in the data, their
  type, and two other additional columns indicating the number of unique
  values and mising values.
}
\description{
\code{get_column_features} is a function to show the type, unique values and
missing values of the column(s) in a dataset
}
\details{
This function creates a tibble object that contains the name of the column(s)
in the dataset along with their class. The number of unique values for each
column(s) and how many missing values there are for each one of them can also
be given.
}
\examples{
get_column_features(
  data = dummy_gasprice
)
}
