% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_get_date_range.R
\name{get_date_range}
\alias{get_date_range}
\title{Get details on start and end dates in a column}
\usage{
get_date_range(data, date_col = "")
}
\arguments{
\item{data}{A data.table, data.frame or tibble object containing the specified
date_col.}

\item{date_col}{A character string indicating the date column in the data.}
}
\value{
A list object containing: \itemize{ \item{start_ym}{ - starting date
 in yyyymm numeric format} \item{start_year}{ - starting year in yyyy numeric
 format} \item{start_month}{ - starting month in mm numeric format}
 \item{end_ym}{ - ending date in yyyymm numeric format } \item{end_year}{ -
 eending year in yyyy numeric format } \item{end_month}{ - eending month in
 mm numeric format} }
}
\description{
\code{get_date_range} is a function to summarize the range of dates in a given
column.
}
\examples{
get_date_range(
  data = dummy_gasprice,
  date_col = "year_month"
)
}
\seealso{
Other Date: 
\code{\link{date_to_period}()},
\code{\link{fast_read_RData_from_network}()},
\code{\link{months_between_periods}()},
\code{\link{parse_date_from_vector}()},
\code{\link{period_to_first_day}()},
\code{\link{period_to_quarter}()}
}
\concept{Date}
