% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_date_vector}
\alias{get_date_vector}
\title{Compute the Period Vector representation of a Decimal Time value}
\usage{
get_date_vector(dtime, frq)
}
\arguments{
\item{dtime}{numeric decimal time value denoting a point in time}

\item{frq}{integer frequency}
}
\description{
The period value will be rounded down to the nearest integer.
This function is not vectorized so only a single value can be
converted at a time.
}
