% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MISC_get_plot_colors.R
\name{get_plot_colors}
\alias{get_plot_colors}
\title{Get vector of plot colors}
\usage{
get_plot_colors(n_colors)
}
\arguments{
\item{n_colors}{A positive integer value indicating the number of color codes
to return.}
}
\value{
A vector of character strings corresponding to ING color codes
}
\description{
\code{get_plot_colors} is a function to generate a vector with color
codes, which can be used for plotting. Currently, only 23 distinct colors are
implemented. Based on recommendations in \url{https://sashat.me/2017/01/11/list-of-20-simple-distinct-colors/}
}
\examples{
get_plot_colors(3)
}
