% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIERARCHICAL_get_ultimate_leaves.R
\name{get_ultimate_leaves}
\alias{get_ultimate_leaves}
\title{Get ultimate leaves}
\usage{
get_ultimate_leaves(data, hierarchical_col, grouping)
}
\arguments{
\item{data}{A tibble containing information on the hierarchy structure. There
should be three variables: the hierarchy groups, their direct parent
("parent_") and whether they are a leaf or not ("leaf_"). The assumption is
that this input is the output of the \code{add_hierarchical_lineage}
function.}

\item{grouping}{A character string indicating for which grouping the ultimate
leaves should be returned.}

\item{hierchical_col}{A character string indicating which column corresponds
to the hierarchy group for which the lineage should be added to the data.}
}
\value{
A vector of character strings corresponding to the ultimate leaves
  for the specified grouping.
}
\description{
\code{get_ultimate_leaves} is a function that creates a vector of the
ultimate leaves corresponding to one of the columns within a specific
grouping, based on data containing the hierarchy information.
}
\examples{
dummy_hierarchical_gasprice \%>\%
   dplyr::select(location, oil_company, level_location, level_oil_company) \%>\%
   dplyr::distinct() \%>\%
   add_hierarchical_lineage(hierarchical_col = "location") \%>\%
   tstools::add_grouping_column(group_cols = c("location", "oil_company")) \%>\%
   get_ultimate_leaves(
      hierarchical_col = "location",
      grouping = "location = New York   &   oil_company = CompanyC"
   )
}
