% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MISC_get_user_corporate_key.R
\name{get_user_corporate_key}
\alias{get_user_corporate_key}
\title{Get user corporate key}
\usage{
get_user_corporate_key()
}
\value{
A character string corresponding to the corporate key of the user
}
\description{
\code{get_user_corporate_key} is a function to determine the corporate key of
the current user of the machine on which the code is being run, based on
system environment variables, user info or a system call. If the corporate
key can't be determined, then UNKNOWN is returned.
}
\examples{
get_user_corporate_key()
}
