% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCiLegendColors}
\alias{getCiLegendColors}
\title{Helper to calculate ci colors for legends}
\usage{
getCiLegendColors(color, n = 1, alpha = NULL)
}
\arguments{
\item{color}{The color of the ci band}

\item{n}{The number if ci bands}

\item{alpha}{The alpha/transparency of the ci band}
}
\value{
A vector of non-transparent colors that result from
oberlaying color over pure white 1:n times
}
\description{
Helper to calculate ci colors for legends
}
\details{
Color may be specified as either a named color or a hex value
Transparency may be specified as a hex value, number 0-255 or number 0-1
}
