% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaaft-surrogates.R
\name{iaaft}
\alias{iaaft}
\title{Generate an iterated amplitude-adjusted Fourier transform surrogate.}
\usage{
iaaft(series, n.max.iter = 150)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}

\item{n.max.iter}{The maximum number of refinement.}
}
\value{
An iaaft representation of the original series, sticking to
  the original values of the series (but otherwise shuffled) and
  preserving the autocorrelation function of the original series.
}
\description{
Generate an iterated amplitude-adjusted Fourier transform surrogate.
}
