% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-iterated-aaft.R
\name{iaaft_surrogate}
\alias{iaaft_surrogate}
\title{Generate an iterated amplitude-adjusted Fourier transform surrogate.
Constrained to the values of the original time series.}
\usage{
iaaft_surrogate(series, n.max.iter = 150)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}

\item{n.max.iter}{The maximum number of refinement.}
}
\value{
An iaaft representation of the original series, sticking to
  the original values of the series (but otherwise shuffled) and
  preserving the autocorrelation function of the original series.
}
\description{
Generate an iterated amplitude-adjusted Fourier transform surrogate.
Constrained to the values of the original time series.
}
