% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTimeSeries.R
\name{importTimeSeries}
\alias{importTimeSeries}
\title{importTimeSeries}
\usage{
importTimeSeries(file, format = c("csv", "xlsx", "json", "zip"))
}
\arguments{
\item{file}{Path to the file to be read}

\item{format}{Which file format is the data stored in? If no format is supplied, importTimeSeries will attempt to guess
from the file extension.}
}
\value{
A named list of ts objects
}
\description{
Import time series data from a file.
If importing from a zip file, the archive should contain a single file with the extension .csv, .xlsx or .json.
}
