% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_indicate_empty_character_cols.R
\name{indicate_empty_character_cols}
\alias{indicate_empty_character_cols}
\title{Replace all empty character column values with an indicator}
\usage{
indicate_empty_character_cols(data, fill_with = "_is_empty")
}
\arguments{
\item{data}{A data.table, data.frame or tibble object containing the data to
which to replace the missing values with the specified indicator.}

\item{fill_with}{A character string indicating what should the mising values
be replaced with.}
}
\value{
A tibble containing the specified character string instead of the
  missing values.
}
\description{
\code{indicate_empty_character_cols} is a function to replace all missing
values in the column(s) of a dataset with the specified indicator.
}
\details{
The function replaces all missing values in the column(s) of a dataset with
the specified indicator.
}
\examples{
indicate_empty_character_cols(
  data = dummy_gasprice,
  fill_with = "_is_empty"
)
}
