% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDefaultTheme.R
\name{initDefaultTheme}
\alias{initDefaultTheme}
\title{Default Configuration List}
\usage{
initDefaultTheme(date_range = NULL)
}
\description{
Create a list containing with the default theme information. 
In case you want to define a custom definition the output of this function is 
a nice starting point.
}

