% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDefaultTheme.R
\name{initDefaultTsTheme}
\alias{initDefaultTsTheme}
\title{Default Configuration List}
\usage{
initDefaultTsTheme()
}
\arguments{
\item{date_range}{character defaults to NULL}
}
\description{
Create a list containing with the default theme information. 
In case you want to define a custom definition the output of this function is 
a nice starting point. This function returns a list containing defaults for
the following plot parameters: 
\describe{
\item{xlim}{range of the date column}
\item{ygrid}{vector of horizontal grid lines}
\item{xlab}{X-axis label}
\item{ylab}{Y-axis label}
\item{lty}{vector of line types}
\item{xaxs}{}
\item{yaxs}{}
\item{xaxt}{}
\item{yaxt}{}
\item{tcl_1}{tick mark length for quarterly ticks}
\item{tcl_2}{tick mark length for yearly ticks}
\item{padj_1}{how far are the labels away from quarterly ticks}
\item{padj_2}{how far are the labels away from yearly ticks}
\item{title_adj}{}
\item{title_line}{}
\item{subtitle_line}{}
\item{title_cex.main}{}
\item{subtitle_cex.main}{}
\item{lwd_ticks_1}{tick mark width for quarterly ticks}
\item{lwd_ticks_2}{tick mark width for yearly ticks}
\item{ygrid_lwd}{line width of the horizontal grid}
\item{ygrid_lty}{line type of the horizontal grid}
\item{ygrid_color}{color of the horizontal grid}
\item{line_colors}{Default colors of the KOF Swiss Economic Institute.}
\item{highlight_window_color}{Background color a highlight window.}
\item{par}{bottom,left,top,right margins}
\item{$height}{}
\item{$width}{}
\item{pointsize}{}
\item{lgnd_cex_label}{}
\item{lgnd_bty}{}
\item{lgnd_vertical_spacing}{}
\item{lgnd_inset}{}
\item{lgnd_xpd}{}
\item{fillUpPeriod}{}
}
}

