% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_initialize_ts_forecast_data.R
\name{initialize_ts_forecast_data}
\alias{initialize_ts_forecast_data}
\title{Initialize time series forecast data}
\usage{
initialize_ts_forecast_data(
  data,
  date_col = "",
  col_of_interest = "",
  group_cols = NULL,
  xreg_cols = NULL,
  hierarchical_cols = NULL
)
}
\arguments{
\item{data}{A data.table, data.frame or tibble object containing the data to
be used for the forecasting.}

\item{date_col}{A character string indicating the date column in the data,
which can also be in period (YYYYMM) format.}

\item{col_of_interest}{A character string indicating the column of interest
in the data to be forecasted.}

\item{group_cols}{A character string or vector of character strings
indicating the column(s) in the data by which to group. The specified
group_cols will be used to create a new column called 'grouping', which
contains every available combination of the specified group_cols (see
example). When continuing with the time series forecasting, every grouping
will be considered separately in both model estimation as well as
performance evaluation.}

\item{xreg_cols}{A character string or vector of character strings indicating
the column(s) in the data to use as external regressors when forecasting.
When xreg_cols have been specified, multivariate time series forecasting
will be performed (instead of univariate time series forecasting if no
xreg_cols are specified).}

\item{hierarchical_cols}{A character string or vector of character strings
indicating the columns in the data that give the numeric hierarchical
information about the tree structure of the available levels. Note that the
levels must be in numeric form with increments of one between adjacent
levels. Specify this parameter if you want to do hierarchical time series
forecasting. Also note that there must be a column named
"level_<hierarchical_col>" for each hierarchical column indicated.}
}
\value{
A tibble containing several columns of data required for forecasting
  within the time series forecasting framework.
}
\description{
\code{initialize_ts_forecast_data} is a function to prepare any dataset for
use within the time series forecasting framework of the tsforecast
package. The function renames a selection of indicated columns to a global
format that is used within the time series forecasting framework. Currently
only datasets with a monthly time frequency are supported.
}
\examples{
initialize_ts_forecast_data(
  data = dummy_gasprice,
  date_col = "year_month",
  col_of_interest = "gasprice",
  group_cols = c("state", "oil_company"),
  xreg_cols = c("spotprice", "gemprice")
)
}
