% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests-helper-functions.R
\name{is_valid_input}
\alias{is_valid_input}
\title{Checks if an input vector is valid.}
\usage{
is_valid_input(x)
}
\arguments{
\item{x}{A vector to check.}
}
\description{
Checks if an input vector is valid.
}
