% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotFunctions.R
\name{karea}
\alias{karea}
\title{Stacked area chart}
\usage{
karea(x, ylab = NULL, title = NULL)
}
\arguments{
\item{x}{object of class multivariate time series}

\item{ylab}{title for the y-axis, NULL by default}

\item{title}{title in the plot, NULL by default}
}
\description{
Stacks the area chart for each column in the multivariate time series
}
\examples{
ts_object <- ts(matrix(rnorm(48),nrow=24),start=c(2008,1), frequency=12)
karea(ts_object) 
}
\author{
Florian Eckert and Heiner Mikosch, KOF, ETH Zurich
}

