% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOF.R
\docType{data}
\name{KOF}
\alias{KOF}
\title{KOF Barometer - Swiss Business Cycle Indicator}
\format{A list of two time series of class ts
\describe{
  \item{barometer}{KOF Barometer Indicator}
  \item{ch_gdp_grwth}{Swiss GDP growth}
  ...
}}
\source{
\url{http://kof.ethz.ch/}
}
\usage{
KOF
}
\description{
A list of time series containing two time series the KOF Barometer and the growth of Swiss GDP over time. KOF Barometer is a monthly business cycle indicator computed by the KOF Swiss Economic Institute. The GDP growth rate is used as a reference series to the Barometer.
}
\keyword{datasets}

