% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_to_q.R
\name{m_to_q}
\alias{m_to_q}
\title{Turn monthly series with regular NAs to quarter}
\usage{
m_to_q(series)
}
\arguments{
\item{series}{an object of class ts with monthly frequency}
}
\description{
Monthly series with NAs in non-quarter months are turned to quarterly series.
Series without NAs are just returned.
}
