% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_months_between_periods.R
\name{months_between_periods}
\alias{months_between_periods}
\title{Determine number of months between two periods}
\usage{
months_between_periods(start, end)
}
\arguments{
\item{start}{The starting period in numeric yyyymm format.}

\item{end}{The ending period in numeric yyyymm format.}
}
\value{
A number indicating the number of months between the start and end
  periods.
}
\description{
\code{months_between_periods} is a function to count the number of months
between two periods.
}
\examples{
months_between_periods(start = 200901, end = 201512)
}
\seealso{
Other Date: 
\code{\link{date_to_period}()},
\code{\link{fast_read_RData_from_network}()},
\code{\link{get_date_range}()},
\code{\link{parse_date_from_vector}()},
\code{\link{period_to_first_day}()},
\code{\link{period_to_quarter}()}
}
\concept{Date}
