\name{nemenyi}
\alias{nemenyi}

\title{Friedman and Nemenyi tests}

\description{This function has been moved to greybox package.}

\usage{
nemenyi(...)
}

\arguments{
  \item{...}{
    Parameters for nothing...
}
}
\value{
This function does not return anything anymore.
}
\references{
Hibbon M., Crone S. F., Kourentzes N., 2012. Statistical significance of forecasting methods: An empirical evaluation of the robustness and interpretability of the MCB, ANOM and Nemenyi tests, The 32nd Annual International Symposium on Forecasting, Boston.
}
\author{
Ivan Svetunkov and Nikolaos Kourentzes
}
\examples{
# No examples here...
}
\keyword{ ~friedman.test }