% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomes.R
\name{norm.square}
\alias{norm.square}
\title{Row-wise sum of squares normalization of a matrix}
\usage{
norm.square(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
a matrix of normalized values
}
\description{
Row-wise sum of squares normalization of a matrix
}

