% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomes.R
\name{norm.sum}
\alias{norm.sum}
\title{Row-wise sum of row normalization of a matrix}
\usage{
norm.sum(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
a matrix of normalized values
}
\description{
Row-wise sum of row normalization of a matrix
}
