% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUMERIC_numeric_cols_to_character.R
\name{numeric_cols_to_character}
\alias{numeric_cols_to_character}
\title{Make all (non-excluded) numeric columns character}
\usage{
numeric_cols_to_character(data, exclude = c())
}
\arguments{
\item{data}{A data.table, data.frame or tibble object containing the data to
which to transform the numerical column(s) into character.}

\item{exclude}{A character string or vector of character strings indicating
the column(s) in the data to not consider while transforming the numerical
column(s)}
}
\value{
A tibble of the initial dataset but with the numerical column(s)
  transformed into character column(s).
}
\description{
\code{numeric_cols_to_character} is a function to transform the numeric
column(s) into character column(s).
}
\details{
The function finds the numeric column(s) in a dataset, and transform all the
numerical values into character./
}
\examples{
numeric_cols_to_character(
  data = dummy_gasprice,
  exclude = c("gasprice", "spotprice")
)
}
