% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocampo.R
\name{ocampo}
\alias{ocampo}
\title{Get ocampo paramters per instance}
\usage{
ocampo(coverage = coverage, references = references, beforeRef = 200,
  afterRef = 1000, smoothingWindow = 100, spacing.low = 130,
  spacing.high = 220, shift.low = -70, shift.high = 60, mc.cores = 2)
}
\description{
Get ocampo paramters per instance
}
