% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-dim-boxcount.R
\name{optimise_dim_boxcount}
\alias{optimise_dim_boxcount}
\title{Finds the minimal embedding dimension of a series 'v' according to
the boxcounting dimension criterion.}
\usage{
optimise_dim_boxcount(v)
}
\arguments{
\item{v}{A numeric vector containing the series.}
}
\value{
The embedding dimension according to the
  boxcounting dimension criterion.
}
\description{
Finds the minimal embedding dimension of a series 'v' according to
the boxcounting dimension criterion.
}
