% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-dim-FNN.R
\name{optimise_dim_FNN}
\alias{optimise_dim_FNN}
\title{Finds the minimal embedding dimension of a series 'v' according to
the method of false nearest neighbours (FNN).}
\usage{
optimise_dim_FNN(v, max.embedding.dimension = 10, embedding.lag = 1,
  orbital.lag = NULL, lag.max = ceiling(length(v) * 0.2),
  lag.method = "correlation", plot.lag.method = F, threshold = 0.95)
}
\arguments{
\item{v}{A numeric vector containing the series.}

\item{max.embedding.dimension}{The maximum embedding dimension to consider}

\item{embedding.lag}{The embedding lag.}

\item{orbital.lag}{The Theiler window. An orbital lag to avoid
temporal correlation. Defaults to NULL, in which case the orbital
lag is chosen as the first local minima of the autocorrelation
function (method = "acf" or method = "autocorrelationfunction")
or the lagged mutual information function (method = "mi" or
method = "mutual information").}

\item{lag.max}{The maximum number of lags for the autocorrelation
function (method = "acf" or method = "autocorrelation function")
or the mutual information function (method = "mi" or
method = "mutual information function").}

\item{lag.method}{The method to compute the orbital lag (Theiler window),
which excludes temporal neighbours to reduce correlation bias.}

\item{plot.temporal.correlation.function}{Should the function used
to determine the orbital lag (temporal separation window) be
plotted?}
}
\value{
The minimum embedding dimension yielding zero false neighbours
  or a false neighbours proportion below some threshold.
}
\description{
The idea is from the following paper:
M. B. Kennel, R. Brown, and H. D. I. Abarbanel,
Determining embedding dimension for phase-space reconstruction
using a geometrical construction, Phys. Rev. A 45, 3403 (1992).
}
\details{
Implemented in the TISEAN package by Reiner Hegger, Holger Kantz
and Thomas Schreiber, and ported to R in the tseriesChaos package
by Antonio Fabio Di Narzo
}
