% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-dim-simplex.R
\name{optimise_dim_simplex}
\alias{optimise_dim_simplex}
\title{Finds the minimal embedding dimension of a series 'v' according to
the method of simplex projection.}
\usage{
optimise_dim_simplex(v, min.embedding.dim = 2, max.embedding.dim = 10,
  embedding.lag = 1, plot.simplex.projection = FALSE)
}
\arguments{
\item{v}{A numeric vector containing the series.}

\item{min.embedding.dim}{The minimum embedding dimension.}

\item{max.embedding.dim}{The minimum embedding dimension.}

\item{embedding.lag}{The embedding lag.}

\item{plot.simplex.projection}{Should the simplex projection results be plotted?}
}
\value{
The embedding dimension that optimises simplex projection of the series.
}
\description{
Finds the minimal embedding dimension of a series 'v' according to
the method of simplex projection.
}
