% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_sorted_lists.R
\name{overlap_sorted_list}
\alias{overlap_sorted_list}
\title{Concat Time Series list wise}
\usage{
overlap_sorted_list(listA, listB)
}
\arguments{
\item{listA}{list of time series}

\item{listB}{list of time series}
}
\description{
Concat overlapping time series list wise. List needs
to be of same length. Takes names of list B.
}
