% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_ts_lists_by_name.R
\name{overlap_ts_lists_by_name}
\alias{overlap_ts_lists_by_name}
\title{Resolve Overlap Listwise, helpful with SA}
\usage{
overlap_ts_lists_by_name(listA, listB, chunkA = "_f4", chunkB = "_f12")
}
\arguments{
\item{listA}{list of time series often of lower frequency}

\item{listB}{list of time series often of higher frequency}

\item{chunkA}{character chunk representing frequencies, defaults to _f4.}

\item{chunkB}{character chunk representing frequences, defaults to _f12.}
}
\description{
Resolve Overlap Listwise, helpful with SA
}
