% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapSortedLists.R
\name{overlapSortedLists}
\alias{overlapSortedLists}
\title{Concat Time Series list wise}
\usage{
overlapSortedLists(listA, listB)
}
\arguments{
\item{listA}{list of time series}

\item{listB}{list of time series}
}
\description{
Concat overlapping time series list wise. List needs
to be of same length. Takes names of list B.
}

