% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_parse_date_columns.R
\name{parse_date_columns}
\alias{parse_date_columns}
\title{Parse the date column of a data frame into YYYY-MM-DD format}
\usage{
parse_date_columns(data)
}
\arguments{
\item{data}{A data.table, data.frame or tibble object containing the data to
which to get the column features.}
}
\value{
The input data, with the identified Date columns now converted into
  YYYY-MM-DD POSIXct vectors
}
\description{
\code{parse_date_columns} is a function that parses columns that are detected
to be of type Date and parses them into YYYY-MM-DD POISXct format. The
detection is done via the \code{get_feature_with_date_columns} function.
}
\examples{
parse_date_columns(
  data = dummy_gasprice
)
}
