% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_parse_date_from_vector.R
\name{parse_date_from_vector}
\alias{parse_date_from_vector}
\title{Attempt to parse a vector into date format}
\usage{
parse_date_from_vector(input)
}
\arguments{
\item{input}{A vector of input values, transformable to date objects}
}
\value{
A vector of date objects
}
\description{
\code{parse_date_from_vector} is a function to transform an input vector into
a vector of date objects of the format YYYY-MM-DD.
}
\examples{
parse_date_from_vector("10-20-2014")
}
\seealso{
Other Date: 
\code{\link{date_to_period}()},
\code{\link{fast_read_RData_from_network}()},
\code{\link{get_date_range}()},
\code{\link{months_between_periods}()},
\code{\link{period_to_first_day}()},
\code{\link{period_to_quarter}()}
}
\concept{Date}
