% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctFunctions.R
\name{pct}
\alias{pct}
\title{Percentage functions}
\usage{
pct(x)
}
\arguments{
\item{x}{object of class time series}
}
\description{
Computes the percentage change at each time point.
}
\examples{
ts_1 <- ts(rnorm(12), frequency=4, start=c(2004,1))
pct(ts_1)
}
\author{
Florian Eckert and Heiner Mikosch, KOF, ETH Zurich
}

