% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_period_delta.R
\name{period_delta}
\alias{period_delta}
\title{Add/subtract months from a period}
\usage{
period_delta(period, delta = 0)
}
\arguments{
\item{period}{The period to start with in numeric yyyymm format.}

\item{delta}{A positive or negative integer value indicating the number of
months to add or substract.}
}
\value{
A numeric period in yyyymm format resulting from input period +
  delta.
}
\description{
\code{period_delta} is a function to add or subtract a number of months to or
from a given period.
}
\examples{
period_delta(200901, 2)
period_delta(200901, -2)
}
