% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_period_to_first_day.R
\name{period_to_first_day}
\alias{period_to_first_day}
\title{Convert period to first day of the month}
\usage{
period_to_first_day(period)
}
\arguments{
\item{period}{A period value in yyyymm numeric format.}
}
\value{
A date object, which is the first day of a given month.
}
\description{
\code{period_to_first_day} is a function to transform a period into the date
object for the first day of that month.
}
\examples{
period_to_first_day(201805)
}
\seealso{
Other Date: 
\code{\link{date_to_period}()},
\code{\link{fast_read_RData_from_network}()},
\code{\link{get_date_range}()},
\code{\link{months_between_periods}()},
\code{\link{parse_date_from_vector}()},
\code{\link{period_to_quarter}()}
}
\concept{Date}
