% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_period_to_last_day.R
\name{period_to_last_day}
\alias{period_to_last_day}
\title{Convert period to last day of the month}
\usage{
period_to_last_day(period)
}
\arguments{
\item{period}{A period value in yyyymm numeric format.}
}
\value{
A date object, which is the last day of a given month.
}
\description{
\code{period_to_last_day} is a function to transform a period into the date
object for the last day of that month.
}
\examples{
period_to_last_day(201805)
}
