% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE_period_to_quarter.R
\name{period_to_quarter}
\alias{period_to_quarter}
\title{Convert period to quarter}
\usage{
period_to_quarter(period)
}
\arguments{
\item{period}{A period value in yyyymm numeric format.}
}
\value{
A numeric value, which is the quarter corresponding to the period.
}
\description{
\code{period_to_quarter} is a function to transform a period into the numeric
value for the corresponding quarter (e.g. 1 for January, February and March,
2 for April, May, June, etc.).
}
\examples{
period_to_quarter(201805)
}
\seealso{
Other Date: 
\code{\link{date_to_period}()},
\code{\link{fast_read_RData_from_network}()},
\code{\link{get_date_range}()},
\code{\link{months_between_periods}()},
\code{\link{parse_date_from_vector}()},
\code{\link{period_to_first_day}()}
}
\concept{Date}
