% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Computes a periodogram for a time series. Wrapper around multitaper::spec.}
\usage{
periodogram(ts, frequency = 1, dtUnits = "year", deltat = 1000, nw = 4,
  k = 2 * nw)
}
\arguments{
\item{ts}{A time series of equally spaced data for which to compute the periodogram.}

\item{frequency}{The time series is converted to a stats::ts object before
being fed to multitaper::spec.mtm. dtUnits gives the frequency for
the stats::ts object.}

\item{dtUnits}{The time unit for the time series.}

\item{deltat}{The time series is converted to a stats::ts object before
being fed to multitaper::spec.mtm. deltat gives time increment for ths
stats::ts object in terms of dtUnits}

\item{nw}{nw a positive double precision number, the time-bandwidth parameter.}

\item{k}{k a positive integer, the number of tapers, often 2*nw.}
}
\description{
Computes a periodogram for a time series. Wrapper around multitaper::spec.
}
