% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-phase-randomisation.R
\name{phase_randomised_surrogate}
\alias{phase_randomised_surrogate}
\title{Generate a phase randomised surrogate according to Theiler.}
\usage{
phase_randomised_surrogate(series)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}
}
\description{
Generate a phase randomised surrogate according to Theiler.
}
