% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIERARCHICAL_plot_hierarchy.R
\name{plot_hierarchy}
\alias{plot_hierarchy}
\title{Plot hierarchy}
\usage{
plot_hierarchy(hierarchy, hierarchical_col = "", interact = FALSE)
}
\arguments{
\item{hierarchy}{A named list, containing both the hierarchy matrix as well
as the hierarchy data to be used for constructing the hierarchy
visualization.}

\item{hierarchical_col}{A character string indicating which column
corresponds to the hierarchical grouping to be visualized.}

\item{interact}{Boolean, which is set to TRUE if interactions with the
hierarchical tree visualization should be enabled, or set to FALSE if they
should not.}
}
\value{
A visgraph object which visualizes the hierarchical tree for the
  specified hierarchical_col within the hierarchy data.
}
\description{
\code{plot_hierarchy} is a function to visualize the hierarchical tree for
one of the hierarchical columns available within the hierarchy data. The
hierarchy is visualizes as a tree with a layer of nodes for each separate
level and edges between the inidividual nodes to indicate (hierarchical)
relationships between them.
}
\examples{
initialize_ts_forecast_data(
    data = dummy_hierarchical_gasprice,
    date_col = "year_month",
    col_of_interest = "gasprice",
    group_cols = "currency",
    hierarchical_cols = c("location", "oil_company")
  ) \%>\%
  tstools::create_hierarchy_list() \%>\%
  plot_hierarchy(
    hierarchical_col = "location"
  )
}
