% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{plotProfiles}
\alias{plotProfiles}
\title{Plot coverages along genome}
\usage{
plotProfiles(fstart, fend, fchr, profs, cols = c(), ann = NULL,
  ylabel = "coverage", ylims = list(), txdb, ftitle = NA, collapse = T,
  with.genes.highlited = c(), plot.labels = T, grid = F,
  with.average = F)
}
\arguments{
\item{fstart}{start of the genomic window (bp)}

\item{fend}{end of the genomic window (bp)}

\item{fchr}{chromosome of the genomic window}

\item{profs}{list of coverages to be plotted}

\item{cols}{colors of the profiles}

\item{ann}{a dataframe of annotations}

\item{ylims}{list of ylimits for plotting the coverages. list(c(min, max), c(min, max))}

\item{txdb}{a transcription database used for plotting gene models}

\item{ftitle}{title of the plot}

\item{collapse}{collapse gene models (default = TRUE)}

\item{with.genes.highlited}{vector of gene ids that should be highlighted}

\item{plot.labels}{plot labels lf gene (default = TRUE)}

\item{grid}{plot grid (default = TRUE)}

\item{with.average}{plot average (default = FALSE)}

\item{ylable}{the label of the y axis}
}
\description{
Plot coverages along genome
}
