% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomes.R
\name{plotRasterHeatmap}
\alias{plotRasterHeatmap}
\title{Plot Raster Heatmap of matrix}
\usage{
plotRasterHeatmap(mat)
}
\arguments{
\item{mat}{the matrix to be visualized}
}
\description{
Plot Raster Heatmap of matrix
}

