% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-random.R
\name{random_surrogate}
\alias{random_surrogate}
\title{Generate a surrogate series by randomly shuffling the values
of the original time series.}
\usage{
random_surrogate(series)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}
}
\description{
Generate a surrogate series by randomly shuffling the values
of the original time series.
}
