% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swissdata_meta.R
\name{read_swissdata_meta}
\alias{read_swissdata_meta}
\title{Read Meta Information from a Swissdata type of yaml file}
\usage{
read_swissdata_meta(path)
}
\arguments{
\item{path}{character file path to swissdata yaml file.}
}
\description{
Read yaml meta information from swissdata type of meta information
and return an R list.
}
