\name{referrals}
\alias{referrals}
\title{GP Referrals Made (All specialties)}
\description{UK NHS GP Referrals Made (All specialties). April 2008 -- June 2013.}
\usage{referrals}
\format{Time series data}
\source{UK Department of Health: Monthly Activity Return}

\keyword{datasets}
\examples{
tsdisplay(referrals)
}