% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize.R
\name{regularize}
\alias{regularize}
\title{Turn an Irregular Time Series to a Regular, Yearly ts-Based Series}
\usage{
regularize(x, df = "\%Y")
}
\arguments{
\item{x}{an irregular time series object of class zoo or xts.}

\item{df}{character denoting the interval in date notation. Currently only works for "\%Y".}
}
\description{
Adds missing values to turn an irregular time series into a regular one. This function is currently experimental and only works for irregular with gaps of one year or more.
}
\examples{
ts1 <- rnorm(5)
dv <- c(seq(as.Date("2010-01-01"), length = 3, by="3 years"),
seq(as.Date("2018-01-01"), length = 2, by="2 years"))
library(zoo)
xx <- zoo(ts1,dv)
regularize(xx)
}
