% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaaft-surrogates.R
\name{rescale}
\alias{rescale}
\title{Rank ordering to rescale process to distribution. Rescales
X to have distribution from Y.}
\usage{
rescale(X, Y, method = c("shell"))
}
\arguments{
\item{X}{Series X.}

\item{Y}{Series Y.}

\item{method}{Sorting method.}
}
\value{
The rescaled X series.
}
\description{
Rank ordering to rescale process to distribution. Rescales
X to have distribution from Y.
}
