% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-seasonal.R
\name{seasonal_surrogate}
\alias{seasonal_surrogate}
\title{Generate a seasonal surrogate according to Sugihara et al.}
\usage{
seasonal_surrogate(series, t.period = 1)
}
\arguments{
\item{series}{The series for which to generate a surrogate.}
}
\description{
Generate a seasonal surrogate according to Sugihara et al.
}
