% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_split_grouping_column.R
\name{split_grouping_column}
\alias{split_grouping_column}
\title{Split grouping column back into separate columns}
\usage{
split_grouping_column(data, sep = "   &   ")
}
\arguments{
\item{data}{A data.table, data.frame or tibble object that has a column named
"grouping" indicating the grouping information, which has been added using
the tstools::add_grouping_column function.}

\item{sep}{A character string indicating the separator that was used to
combine the separate elements of the grouping.}
}
\value{
A tibble extended with one or more additional group columns, which
  correspond to a split of the 'grouping' column.
}
\description{
\code{split_grouping_column} is a function to split a single grouping column
back into one or more grouping columns.
}
\examples{
data_with_grouping_column <- tstools::add_grouping_column(
  data = dummy_gasprice,
  group_cols = c("state", "oil_company")
)
split_grouping_column(data = data_with_grouping_column)
}
