% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarChartWithNegValues.R
\name{stackedBarChartsWithNegValues}
\alias{stackedBarChartsWithNegValues}
\title{29122016
Stacked bar charts with positive and negative values and supplementary line plots}
\usage{
stackedBarChartsWithNegValues(c_vect, show_sums_as_line = T, theme = NULL)
}
\arguments{
\item{vect}{matrix}

\item{vect_ts}{object of class time series}
}
\description{
29122016
Stacked bar charts with positive and negative values and supplementary line plots
}
\examples{
vect <- cbind(50+rnorm(36), rnorm(36)+c(-2,2), -20+rnorm(36), 100+rnorm(36))
vect_ts <- ts(vect, frequency=12, start=c(2004,1))
stackedBarChartsWithNegValues(vect_ts)
addLinePlot(vect_ts[,2])
}

