% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Computes a set of basic statistics for a scalar valued vector}
\usage{
stats(v)
}
\arguments{
\item{v}{A scalar valued vector.}
}
\value{
A named vector of statistical parameters of the vector 'v'
}
\description{
Computes a set of basic statistics for a scalar valued vector
}
