% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripLeadingNasFromTs.R
\name{stripLeadingNAsFromTs}
\alias{stripLeadingNAsFromTs}
\title{Strip Leading NAs from a Time Series Object}
\usage{
stripLeadingNAsFromTs(s)
}
\arguments{
\item{s}{an object of class ts.}
}
\description{
Removes NAs to begin with and starts time series index at the first non-NA value.
}

