% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUMERIC_summarize_numeric_value.R
\name{summarize_numeric_value}
\alias{summarize_numeric_value}
\title{Transforms a large numeric value to a string.}
\usage{
summarize_numeric_value(value)
}
\arguments{
\item{value}{The numeric value (or values) that needs to be converted.}
}
\value{
A string of rounded numeric value with the corresponding scale.
}
\description{
\code{summarize_numeric_value} is a function to convert a large numeric value
to thousands, millions, billions or trillions.
}
\examples{
summarize_numeric_value(value = 124401240)
}
