% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm-surrogate.R
\name{surrogate_ccm}
\alias{surrogate_ccm}
\title{Cross-mapping for a surrogate ensemble.}
\usage{
surrogate_ccm(original.data, E = 2, tau = 1,
  library.size = ceiling(nrow(original.data) * 0.9), with.replacement = T,
  RNGseed = 1111, exclusion.radius = E + 1, num.neighbours = E + 1,
  epsilon = NULL, silent = T, lag = 0, lib = c(1,
  nrow(original.data)[1]), pred = lib, num_neighbors = E + 1,
  random.libs = TRUE, library.column = 1, target.column = 2,
  surrogate.column = target.column, samples.surrogates = 100,
  n.surrogates = 100, parallel = F, surrogate.method = "aaft")
}
\description{
Cross-mapping for a surrogate ensemble.
}
