% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate.R
\name{surrogates}
\alias{surrogates}
\title{Generates surrogate time series for 'ts', which is a vector of ordered scalar observations.}
\usage{
surrogates(ts, surrogate.method, n.surrogates, print.to.console = F)
}
\arguments{
\item{ts}{Time series for which to generate surrogate time series. A vector.}

\item{n.surrogates}{The number of surrogates to generate.}

\item{method}{String indicating the surrogate time series generation method.}
}
\description{
WARNING: Adaptations to the fractal::surrogate function had to be done to actually return,
         the surrogate data (it only plotted the data).
}
\details{
Parameters:
}
