% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcFunctions.R
\name{ts_pred_uniform}
\alias{ts_pred_uniform}
\title{Prediction of uniform temporal disaggregation of time series}
\usage{
ts_pred_uniform(x)
}
\arguments{
\item{x}{multivariate time series object}
}
\description{
Common function to calculate different growth contributions.
}
\examples{
mat <- matrix(c(7+rnorm(48),50+rnorm(48),70+rnorm(48),100+rnorm(48)),ncol=4)
mts_object <- ts(mat, frequency=12, start=c(2004,1))
colnames(mts_object) <- c("IT","DE","CH","FR")
ts_pred_uniform(mts_object)
}
\author{
Florian Eckert and Heiner Mikosch, KOF, ETH Zurich
}

