% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsidx.R
\name{tsidx}
\alias{tsidx}
\title{Get an index}
\usage{
tsidx(x, byear, bperiod = NULL)
}
\arguments{
\item{x}{object of class time series}

\item{byear}{year of observation}

\item{bperiod}{month or quarter of observation, NULL by default (if yearly data). E.g. 2nd quarter.}
}
\description{
Compare time series to extracted value at given time point (year and quarter e.g.).
}
\examples{
ts_1 <- ts(rnorm(20), start=c(2004,1), frequency=4)
tsidx(ts_1,2005,4)
}
\author{
Florian Eckert and Heiner Mikosch, KOF, ETH Zurich
}

