% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Plot time series}
\usage{
tsplot(..., tsr = NULL, left_as_bar = FALSE, group_bar_chart = NULL,
  plot_title = NULL, plot_subtitle = NULL, plot_subtitle_r = NULL,
  find_ticks_function = "findTicks", fill_up_start = FALSE,
  overall_xlim = NULL, overall_ylim = NULL, manual_date_ticks = NULL,
  manual_value_ticks_l = NULL, manual_value_ticks_r = NULL, theme = NULL,
  quiet = TRUE, auto_legend = TRUE)
}
\arguments{
\item{...}{}

\item{tsr}{}

\item{left_as_bar}{}

\item{group_bar_chart}{}

\item{plot_title}{}

\item{plot_subtitle}{}

\item{plot_subtitle_r}{}

\item{find_ticks_function}{}

\item{fill_up_start}{}

\item{overall_xlim}{}

\item{overall_ylim}{}

\item{manual_date_ticks}{}

\item{manual_value_ticks_l}{}

\item{manual_value_ticks_r}{}

\item{theme}{}

\item{quiet}{}

\item{auto_legend}{}
}
\description{
Doc to be expanded later. This just fixes #33.
}
