% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot2y.R
\name{tsplot2y}
\alias{tsplot2y}
\title{Create Time Series Plots With Two Ordinates (Y-axes)}
\usage{
tsplot2y(x, y, ..., right_as_barchart = F, theme_2y = NULL,
  plot.title = NULL, plot.subtitle = NULL, lgnd = NULL, write_pdf = F,
  crop_pdf = F, fname = NULL, highlight_window = NULL, theme_out = F,
  ygrid_factor = 4, l_manual_y_range = NULL, r_manual_y_range = NULL)
}
\arguments{
\item{x}{list of time series, related to the right Y-axis}

\item{y}{list of time series, related to the left Y-axis}

\item{...}{list of additional adhoc plot options.}

\item{plot.title}{character}

\item{plot.subtitle}{character}

\item{theme}{list}
}
\description{
Create Time Series Plots With Two Ordinates (Y-axes)
}

