% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsqm.R
\name{tsqm}
\alias{tsqm}
\title{Interpolate quarterly time series into monthly}
\usage{
tsqm(qts)
}
\arguments{
\item{qts}{quarterly time series}
}
\description{
Repeat quarterly variables two times to generate a monthly variable.
}
\examples{
tsq <- ts(1:20,start=c(1990,1),frequency = 4)
tsqm(tsq)

}
