% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{tstools-deprecated}
\alias{tstools-deprecated}
\alias{computeDecimalTime}
\alias{tstools-deprecated-package}
\title{Deprecated function(s) in tstools}
\usage{
computeDecimalTime(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the tstools package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{computeDecimalTime} \tab now a synonym for \code{\link{compute_decimal_time}}\cr
}
}

