\name{twoSigmas}
\alias{twoSigmas}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
    Prediction intervals based on two standard deviations
}
\description{
    This function returns prediction intervals based on two separate sigmas for left and right sides of the distribution
}
\usage{
    twoSigmas(model, level)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{Object of the class 'smooth'.}
  \item{level}{Confidence level for the intervals construction.}
}
\details{
    Nothing here yet.
}
\value{
  \item{lower}{Lower bound of prediction interval.}
  \item{upper}{Upper bound of prediction interval.}
  \item{sigma}{Matrix with lower and upper sigmas.}
}
\references{
    Nothing here yet.
}
\author{
    Ivan Svetunkov
}

\seealso{
    Nothing
}
\examples{

}
\keyword{ htest }
