% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_unlist_if_required.R
\name{unlist_if_required}
\alias{unlist_if_required}
\title{Extract item from a list of length 1}
\usage{
unlist_if_required(object)
}
\arguments{
\item{object}{List with only one item in it.}
}
\value{
The item in the list.
}
\description{
\code{unlist_if_required} is a function to take an item, usually a data
frame, out of a list of length one.
}
\examples{
data_in_a_list <- list(data = dummy_gasprice)
unlist_if_required(object = data_in_a_list)
}
