% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate-ensemble.R
\name{validate_surrogate_method}
\alias{validate_surrogate_method}
\title{Checks if a given surrogate method is among the currently
implemented surrogate methods.}
\usage{
validate_surrogate_method(surrogate.method)
}
\arguments{
\item{surrogate.method}{The surrogate method to check.}
}
\description{
Checks if a given surrogate method is among the currently
implemented surrogate methods.
}
