\name{vec}
\alias{vec}
\title{Coerce an Object to a Vector}
\description{
  A generic function to coerce an object to a vector.
}
\usage{
vec(x)
}
\arguments{
  \item{x}{An R object.}
}
\value{
  A vector representation of \code{x}.
}
\seealso{
\code{\link{vec.list}}}
\keyword{methods}
\keyword{utilities}

