% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{wide_to_ts}
\alias{wide_to_ts}
\title{Transform a wide format data.frame into a tslist}
\usage{
wide_to_ts(data)
}
\arguments{
\item{data}{data.frame The data.frame to be transformed}
}
\description{
The time series in the data.frame may be stored either rowwise or columnswise.
The identifying column must be called date (for columnwise) or series (for rowwise)
}
