% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts}
\alias{write_ts}
\title{Export a list of time series to a file.}
\usage{
write_ts(tl, fname = "timeseriesdb_export", format = "csv",
  date_format = NULL, timestamp_file = TRUE, round_digits = NULL, ...)
}
\arguments{
\item{tl}{list of time series}

\item{fname}{character file name. If set to NULL a standard file name chunk + Sys.Date is used.}

\item{format}{character denotes export formats. Defaults to .csv. "csv", "xlsx", "json", "rdata" are available. Spreadsheet formats like csv allow for further optional parameters.}

\item{date_format}{character denotes the date format. Defaults to NULL. If set to null the default is used: Jan 2010.}

\item{timestamp_file}{If TRUE, the current date will be appended to the file name}

\item{round_digits}{integer, precision in digits.}

\item{...}{additional arguments used by spedific formats.}
}
\description{
Export a list of time series to a file.
}
\details{
Additional arguments covered by \code{...}
\tabular{lll}{
  \strong{Name} \tab \strong{Effect} \tab \strong{Format(s)} \cr
  \code{wide} \tab Export data in a wide format (one column per series) \tab CSV, XLSX \cr
  \code{transpose} \tab Transpose exported data (one row per series) \tab CSV, XLSX, only if wide = TRUE \cr
  \code{zip} \tab If set to TRUE, the file is compressed into a zip archive after export \tab any \cr
  \code{rdata_varname} \tab The name of the time series list object in the exported RData file \tab rdata \cr
}
}
