% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTslist.R
\name{writeTslist}
\alias{writeTslist}
\alias{writeTslist.toJSON}
\title{EXPERIMENTAL: Write a List of Time Series to .csv, .xlsx, or .json}
\usage{
writeTslist(tsl, fn = NULL, date_format = NULL, ...)

\method{writeTslist}{toJSON}(tsl, fn = NULL, date_format = NULL, ...)
}
\arguments{
\item{tsl}{list of time series}

\item{fn}{character filename, defaults to NULL which sets the file name automatically.}

\item{date_format}{specify an output date format using \code{\link{format}}'s syntax.}

\item{...}{additional optional parameters which may be method specific. the 'wide' parameter is specific to .xlsx and .csv exports.}
}
\description{
The writeTslist generic is usually used wrapped 
within \code{\link{exportTslist}}. It writes a list
of multiple time series objects to disk according
to the class of the list of time series. You may want
to set this class manually and call the writeTsList directly, but using the wrapper for writing is the way to go for most people.
}
