% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcFunctions.R
\name{yoygrc}
\alias{yoygrc}
\title{# Year-over-year Growth Contributions}
\usage{
yoygrc(x)
}
\arguments{
\item{x}{multivariate time series object}
}
\description{
Uses function ts_pred_uniform to predict the uniform temporal disaggregation of time series.
}
\examples{
mat <- matrix(c(7+rnorm(48),50+rnorm(48),70+rnorm(48),100+rnorm(48)),ncol=4)
mts_object <- ts(mat, frequency=12, start=c(2004,1))
colnames(mts_object) <- c("IT","DE","CH","FR")
yoygrc(mts_object)
}
\author{
Florian Eckert and Heiner Mikosch, KOF, ETH Zurich
}

