% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctFunctions.R
\name{yoypct}
\alias{yoypct}
\title{Percentage function}
\usage{
yoypct(x)
}
\arguments{
\item{x}{object of class time series}
}
\description{
Computes the year-over-year percentage change at each time point of the time series.
}
\examples{
ts_1 <- ts(rnorm(12), frequency=4, start=c(2004,1))
yoypct(ts_1)
}
\author{
Florian Eckert and Heiner Mikosch, KOF, ETH Zurich
}

