\name{TStools-defunct}
\alias{TStools-defunct}
\title{Defunct Functions in Package \pkg{TStools}}

\alias{abc}
\alias{abcxyz}
\alias{cmav}
\alias{coxstuart}
\alias{decomp}
\alias{geomean}
\alias{get.opt.k}
\alias{lagmatrix}
\alias{lambdaseq}
\alias{leadtrail}
\alias{nemenyi}
\alias{residout}
\alias{seasdummy}
\alias{seasplot}
\alias{theta}
\alias{trendtest}
\alias{wins}
\alias{xyz}
\alias{elm}
\alias{mlp}
\alias{linscale}
\alias{plot.net}

\description{
    The functions listed here have graduated from \pkg{TStools} to other packages and are now available on CRAN. Execute functions to get a message about their new home package.
}

\usage{
    abc(...)
    abcxyz(...)
    cmav(...)
    coxstuart(...)
    decomp(...)
    geomean(...)
    get.opt.k(...)
    lagmatrix(...)
    lambdaseq(...)
    leadtrail(...)
    nemenyi(...)
    residout(...)
    seasdummy(...)
    seasplot(...)
    theta(...)
    trendtest(...)
    wins(...)
    xyz(...)
    elm(...)
    mlp(...)
    linscale(...)
    plot.net(...)
}

\details{
    Package \pkg{tsutils} has welcomed these functions: abc, abcxyz, cmav, coxstuart, decomp, geomean, get.opt.k (renamed to getOptK), lagmatrix, lambdaseq, leadtrail, nemenyi, residout, seasdummy, seasplot, theta, trendtest, wins, xyz.
    Package \pkg{nnfor} has welcomed these functions: elm, mlp, linscale, plot.net. 
}

\keyword{internal}