% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Bin a dataset with multiple columns by one of the columns.}
\usage{
bin(dt, bin.size, by, bin.average.function = mean.narm, interpolate = T,
  remove.na = T, time.sampled.at = "mid", start = NULL, end = NULL,
  add.binning.info = F)
}
\arguments{
\item{dt}{A data frame containing the data to be binned.}

\item{bin.size}{The size of the bins}

\item{by}{The name of the column to bin by.}

\item{bin.average.function}{The function to use for averaging bins. Default
is the arithmetic mean.}

\item{interpolate}{Should empty bins be interpolated linearly?}

\item{remove.na}{Should NAs remaining after interpolation (usually at
endpoints after interpolatin) be removed? BEWARE: be careful about removing
 nans before interpolating.}

\item{time.sampled.at}{How time indices are constructed. Either 'start',
'mid' or 'end' of bins. Defaults to 'mid'.}

\item{start}{The minimum value of the 'by' column for which to subset data.
Defaults to the first lower bin limit < min(dt[, by]), obeying bin.size.}

\item{end}{The maximum value of the 'by' column for which to subset data.
Defaults to the first upper bin limit > max(dt[, by]), obeying bin.size.}

\item{add.binning.info}{Should info on the binning be added to the resulting
dataframe?}
}
\description{
Bin a dataset with multiple columns by one of the columns.
}
