% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Bin a dataset by a common column "Age".}
\usage{
bin(dt, bin.size, bin.average.function = mean.narm, interpolate = T,
  remove.na = T)
}
\arguments{
\item{dt}{A data frame containing the data to be binned.}

\item{dt}{A data frame containing the data to be binned.}
}
\description{
Bin a dataset by a common column "Age".
}

