% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Bin a dataset by a common column "Age".}
\usage{
bin(dt, bin.size, common.column, bin.average.function = mean.narm,
  interpolate = T, remove.na = T, agepoint = "mid")
}
\arguments{
\item{dt}{A data frame containing the data to be binned.}

\item{bin.size}{The size of the bins}

\item{bin.average.function}{The function to use for averaging bins. Defaults to mean function.}

\item{interpolate}{Should empty bins be interpolated linearly?}

\item{remove.na}{Should NAs remaining after interpolation (usually at endpoints after interpolatin) be removed? BEWARE: be careful about removing nans before interpolating.}

\item{agepoint}{How should time indices be constructed? Start, mid or endpoint of bins?}
}
\description{
Bin a dataset by a common column "Age".
}
