% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm-over-library-sizes.R
\name{ccm_over_library_sizes}
\alias{ccm_over_library_sizes}
\title{Performs CCM over multiple library sizes. This function only exists to allow
parallelisation over library sizes at the lowermost level.}
\usage{
ccm_over_library_sizes(lag, data, E = 2, tau = 1, library.sizes = 100,
  low.libsize = min(E * tau + max(2, abs(lag)), 10, 20, na.rm = T),
  n.libsizes.to.check = 30, high.libsize = max(library.sizes), lib = c(1,
  dim(data)[1]), pred = lib, samples.original = 100,
  samples.surrogates = 0, n.surrogates = 0, surrogate.method = "AAFT",
  time.unit = "bins", time.bin.size = 1, num.neighbours = E + 1,
  random.libs = TRUE, with.replacement = TRUE, exclusion.radius = E,
  epsilon = NULL, RNGseed = 1111, parallel = F, time.run = F,
  print.to.console = F, time.series.length.threshold = 100,
  library.column = 1, target.column = 1, surrogate.column = target.column,
  silent = T)
}
\arguments{
\item{lag}{The lag (called prediction horizon in rEDM::ccm) for which to
compute CCM.}

\item{data}{A data frame containing two columns - one for the presumed driver
and one for the response.}

\item{E}{The embedding dimension. Defaults to NULL, which triggers automated
optimisation of the embedding dimension up to the dimension specified by
'max.E'.}

\item{tau}{The embedding lag. Defaults to NULL, which triggers automated
optimisation of the embedding lag up to the dimension specified by 'max.tau'.
 For sparsely sampled time series (for example geological time series), it is
  wise to set this value to 1. For densely sampled time series, this should
  be set to the first minima of the autocorrelation function of the presumed
  driver.}

\item{library.sizes}{Either a single maximum library size (cross mapping
is performed for a range of value from the smallest possible library
size to the provided library size) or a user-specified range
of library sizes. If user-provided, make sure to provide at least
20 different library sizes to ensure robust convergence assessment.}

\item{low.libsize}{If one library size is specified, cross map
for library sizes ranging from 'low.libsize' to 'high.libsize'.}

\item{n.libsizes.to.check}{Minimum number of library sizes for the
convergence test.}

\item{high.libsize}{If one library size is specified, cross map
  for library sizes ranging from 'low.libsize' to 'high.libsize'.
#' @param data A data frame containing two columns - one for the presumed driver
  and one for the response.}

\item{lib}{Indices of the original library time series to use as the library
(training) set.}

\item{pred}{Indices of the original target time series to use as prediction
set. If this overlaps with the training set, make sure to use leave-K-out
cross validation setting the 'exclusion.radius' parameters to a minimum of
E + 1.}

\item{samples.original}{The number of random libraries to draw when
calculating the cross map skill.}

\item{samples.surrogates}{The number of surrogate time series in the null
ensemble.}

\item{n.surrogates}{Should a surrogate test also be performed? If so, 'n.surrogates' sets
 the number of surrogate time series to use. By default, no surrogate test is performed
(n.surrogates = 0).}

\item{surrogate.method}{Which method should be used to generate surrogate
time series? Defaults to "AAFT". For more options, see the description of
the 'surrogate_ensemble' function in this package.}

\item{time.unit}{The time unit of the raw time series.}

\item{time.bin.size}{The temporal resolution of the raw time series (given in the units
indicated by 'time.unit').}

\item{num.neighbours}{The number of nearest neighbours to use in predictions.
Defaults to E + 1.}

\item{random.libs}{Whether or not to sample random library (training) sets.
Defaults to TRUE.}

\item{with.replacement}{Should samples be drawn with replacement? Defaults
to TRUE.}

\item{exclusion.radius}{The number of temporal neighbours to exclude for the
leave-K-out cross validation. Defaults to E + 1.}

\item{epsilon}{Exlude neighbours if the are within a distance of 'epsilon'
from the predictee.}

\item{RNGseed}{A random number seed. For reproducibility.}

\item{parallel}{Activate parallellisation? Defaults to true. Currently,
this only works decently on Mac and Linux systems.}

\item{time.run}{Time the run?}

\item{print.to.console}{Display progress?}

\item{time.series.length.threshold}{Display a warning if the time series length drops
below this threshold.}

\item{library.column}{Integer indicating which column to use as the library
column (presumed response).}

\item{target.column}{Integer indicating which column to use as the target
column (presumed driver). Defaults to the opposite of 'library.column'.}

\item{surrogate.column}{Which column to use to generate surrogates. Defaults
to the value of 'target.column' (the presumed driver).}

\item{silent}{Suppress warnings?}
}
\description{
Performs CCM over multiple library sizes. This function only exists to allow
parallelisation over library sizes at the lowermost level.
}
