% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{coverageWindowsStranded}
\alias{coverageWindowsStranded}
\title{Get windowed strand-oriented coverages}
\usage{
coverageWindowsStranded(windows, coverage)
}
\arguments{
\item{windows}{a data frame of windows with columns 'chr', 'start', 'end', 'strand'}

\item{coverage}{a coverage object (\code{\link[IRanges]{RleList}} as returned by \code{\link[IRanges]{coverage}})}
}
\value{
a matrix
}
\description{
Out-of-bound windows will be removed!
}
