% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-random-ages.R
\name{draw_random_ages}
\alias{draw_random_ages}
\title{A wrapper around agemodel that returns a matrix of multiple age models.
Each column is an age model.}
\usage{
draw_random_ages(ages, sigmas, n.replicates = 1, n.sigma = 2,
  firstdiffagreementratio = 1, tolerance = 10^-(10))
}
\arguments{
\item{ages}{A vector containing the age data.}

\item{sigmas}{A vector containing the 1 sigma uncertainties associated with
the age data.}

\item{n.replicates}{How many age vectors to draw?}

\item{n.sigma}{n*sigma uncertainty. Defaults to 2.}

\item{firstdiffagreementratio}{How large can the slope between time steps be?
Defaults to 1, or 100 percent.}

\item{tolerance}{A tolerance level to speed up computations when age points
are very close. Ensures that iterations don't get stuck.}
}
\description{
A wrapper around agemodel that returns a matrix of multiple age models.
Each column is an age model.
}
