% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-local-minima.R
\name{first_local_minima}
\alias{first_local_minima}
\title{Finds the index of the first local minima in a series.
If value == F and no local minima is found, returns NA.
If value == T and no local minima is found, returns min(v)}
\usage{
first_local_minima(v, value = F)
}
\arguments{
\item{v}{A numeric vector containing the series.}

\item{value}{Return the value at the local minima instead of the index?}
}
\value{
The index or value of the first local minima.
}
\description{
Finds the index of the first local minima in a series.
If value == F and no local minima is found, returns NA.
If value == T and no local minima is found, returns min(v)
}
