% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm-convergence-checks.R
\name{get_convergence_parameters}
\alias{get_convergence_parameters}
\title{Computes various parameters that can be used to determine if
cross-map skill increases with increasing library size.}
\usage{
get_convergence_parameters(ccm.result, confidence.level = 0.99, plot = F)
}
\arguments{
\item{ccm.result}{A data frame}
}
\description{
Computes various parameters that can be used to determine if
cross-map skill increases with increasing library size.
}
