% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaaft-surrogates.R
\name{iAAFT}
\alias{iAAFT}
\title{Iteratively generate a surrogate series with the same spectrum and distribution
as the original series (iterated amplitude-adjusted Fourier transform, or iAAFT).
Uses Fourier phase shuffling on Xcorrelogram and rank ordering to retain the
distribution of Xdistribution.}
\usage{
iAAFT(Xcor, Xdist = Xcor, tolerance = 0.01, maxit = 100,
  adjust.var = TRUE, zero.mean = TRUE, quiet = TRUE, diff = FALSE,
  criterion = c("periodogram"), rel.convergence = TRUE,
  method = c("shell"))
}
\arguments{
\item{Xcor}{Record whose periodogram should be kept}

\item{Xdist}{Record whos distribution should be kept}

\item{tolerance}{Convergence criterion}

\item{maxit}{Maximal numbers of iteration}

\item{adjust.var}{Set var(Xcor) to var(Xdist)}

\item{zero.mean}{Set the mean of Xdist to sero and add it later}

\item{quiet}{Forces silence}

\item{diff}{Creates a difference vector for every iteration}

\item{criterion}{"periodogram" or "acf" as criterion for convergence}

\item{rel.convergence}{Use a relative convergence criterion}

\item{method}{Sorting method to use}
}
\value{
An iterated AAFT surrogate series.
}
\description{
Iteratively generate a surrogate series with the same spectrum and distribution
as the original series (iterated amplitude-adjusted Fourier transform, or iAAFT).
Uses Fourier phase shuffling on Xcorrelogram and rank ordering to retain the
distribution of Xdistribution.
}
