% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{long_to_ts}
\alias{long_to_ts}
\title{Transform a long format data.frame of time series to a tslist}
\usage{
long_to_ts(data)
}
\arguments{
\item{data}{data.frame The data.frame to be transformed}
}
\description{
The data.frame must have three columns "date", "value" and "series" (identifying the time series)
}
