% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Computes a periodogram for a time series. Wrapper around multitaper::spec.}
\usage{
periodogram(ts, frequency = 1, dtUnits = "year", deltat = 1000, nw = 4,
  k = 2 * nw)
}
\arguments{
\item{ts}{A time series of equally spaced data for which to compute the periodogram.}

\item{nw}{nw a positive double precision number, the time-bandwidth parameter.}

\item{k}{k a positive integer, the number of tapers, often 2*nw.}
}
\description{
Computes a periodogram for a time series. Wrapper around multitaper::spec.
}
