% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swissdata_meta.R
\name{read_swissdata_meta}
\alias{read_swissdata_meta}
\title{Read swissdata style yaml timeseries metadata}
\usage{
read_swissdata_meta(path, locale = "de", as_list = FALSE)
}
\arguments{
\item{path}{Path to the yaml file to be read}

\item{locale}{Locale in which to read the data (supported are "de", "fr", "it" and "en")}

\item{as_list}{Should the output be converted to a list?}
}
\description{
read_swissdata_meta reads the given .yaml file and converts it into a
per-timeseries format.
}
\details{
If as_list is set to TRUE, the function returns a nested list with one
element per timeseries, otherwise a data.table with one row per series.
}
