% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsggplotly.R
\name{tsggplotly}
\alias{tsggplotly}
\title{Convert ggplot2 time series object to plotly object}
\usage{
tsggplotly(p, ...)
}
\arguments{
\item{p}{ggplot2 figure, as returned by \code{\link{tsggplot}}}

\item{...}{additional arguments passed on to \code{\link[plotly]{ggplotly}}}
}
\description{
Convert ggplot2 time series object to plotly object
}
