% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsggplotly.R
\name{tsggplotly}
\alias{tsggplotly}
\title{Convert ggplot2 time series object to plotly object}
\usage{
tsggplotly(p, ...)
}
\arguments{
\item{p}{ggplot2 figure}

\item{...}{additional arguments}
}
\description{
Convert ggplot2 time series object to plotly object
}
