% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\alias{tsplot.list}
\alias{tsplot.mts}
\alias{tsplot.ts}
\title{Basic Times Series Plot Method}
\usage{
tsplot(series, ..., theme = NULL, plot.title = NULL, plot.subtitle = NULL,
  lgnd = NULL, write_pdf = F, crop_pdf = F, cex_label = 0.65,
  fname = NULL, ygrid = T, ygrid_factor = 4, yaxis_factor, theme_out,
  print_x_axis, print_y_axis = T, print_y_right = F,
  highlight_window = NULL, manual_date_range = NULL,
  manual_value_range = NULL, auto_name = NULL)

\method{tsplot}{ts}(series, ..., theme = NULL, plot.title = NULL,
  plot.subtitle = NULL, lgnd = NULL, write_pdf = F, crop_pdf = F,
  cex_label = 0.65, fname = NULL, ygrid = T, ygrid_factor = 4,
  yaxis_factor = 20, quarter_ticks = T, theme_out = F, print_x_axis = T,
  print_y_axis = T, print_y_right = F, highlight_window = NULL,
  manual_date_range = NULL, manual_value_range = NULL, auto_name = NULL)

\method{tsplot}{mts}(series, ..., theme = NULL, plot.title = NULL,
  plot.subtitle = NULL, lgnd = NULL, write_pdf = F, crop_pdf = F,
  cex_label = 0.65, fname = NULL, ygrid = T, ygrid_factor = 4,
  yaxis_factor = 20, quarter_ticks = T, theme_out = F, print_x_axis = T,
  print_y_axis = T, print_y_right = F, highlight_window = NULL,
  manual_date_range = NULL, manual_value_range = NULL, auto_name = NULL)

\method{tsplot}{list}(series, theme = NULL, plot.title = NULL,
  plot.subtitle = NULL, lgnd = NULL, write_pdf = F, crop_pdf = F,
  fname = NULL, cex_label = 0.65, ygrid = T, ygrid_factor = 5,
  yaxis_factor = 20, quarter_ticks = T, theme_out = F, print_x_axis = T,
  print_y_axis = T, print_y_right = F, highlight_window = NULL,
  manual_date_range = NULL, manual_value_range = NULL, auto_name = NULL,
  ...)
}
\arguments{
\item{series}{object of class ts, or list of time series series.}

\item{...}{objects of class ts}

\item{theme}{list holding extra style arguments to be passed to plot.}

\item{ygrid_factor}{numeric, defaults to 5. Factor by which the difference 
between the maximum and the minimum y value should be divided. This parameter
determines the number of horizontal grid lines when dynamic ygrid_dynamic is 
TRUE.}

\item{ygrid_dynamic}{logical, defaults to FALSE. Should ygrids be dynamic?
If not, the ygrid configuration of the theme is used.}
}
\description{
Basic Times Series Plot Method
}

