% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot2y.R
\name{tsplot2y}
\alias{tsplot2y}
\title{Create Time Series Plots With Two Ordinates (Y-axes)}
\usage{
tsplot2y(x, y, ..., theme = NULL, plot.title = NULL, plot.subtitle = NULL,
  l_manual_y_range = NULL, r_manual_y_range = NULL)
}
\arguments{
\item{x}{list of time series, related to the right Y-axis}

\item{y}{list of time series, related to the left Y-axis}

\item{...}{list of additional adhoc plot options.}

\item{theme}{list}

\item{plot.title}{character}

\item{plot.subtitle}{character}
}
\description{
Create Time Series Plots With Two Ordinates (Y-axes)
}

