% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{tstools-deprecated}
\alias{tstools-deprecated}
\alias{computeDecimalTime}
\alias{concatTs}
\alias{fillupYearWitnNAs}
\alias{importTimeSeries}
\alias{initDefaultTheme}
\alias{overlapSortedLists}
\alias{overlapTslByName}
\alias{resolveOverlap}
\alias{stripLeadingNAsFromTs}
\alias{stripTrailingNAsFromTs}
\alias{writeTimeSeries}
\title{Deprecated function(s) in tstools}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the tstools package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{computeDecimalTime} \tab now a synonym for \code{\link{compute_decimal_time}}\cr
  \code{concatTs} \tab now a synonym for \code{\link{concat_ts}}\cr
  \code{fillupYearWitnNAs} \tab now a synonym for \code{\link{fill_year_with_nas}}\cr
  \code{importTimeSeries} \tab now a synonym for \code{\link{read_ts}}\cr
  \code{init_tsplot_theme} \tab now a synonym for \code{\link{init_tsplot_theme}}\cr
  \code{overlapSortedLists} \tab now a synonym for \code{\link{overlap_sorted_ts_lists}}\cr
  \code{overlapTslByName} \tab now a synonym for \code{\link{overlap_ts_lists_by_name}}\cr
  \code{resolveOverlap} \tab now a synonym for \code{\link{resolve_ts_overlap}}\cr
  \code{stripLeadingNAsFromTs} \tab now a synonym for \code{\link{strip_ts_of_leading_nas}}\cr
  \code{stripTrailingNAsFromTs} \tab now a synonym for \code{\link{strip_ts_of_trailing_nas}}\cr
  \code{writeTimeSeries} \tab now a synonym for \code{\link{write_ts}}\cr
}
}

