% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTimeseries.R
\name{writeTimeSeries}
\alias{writeTimeSeries}
\title{TODO: Dox}
\usage{
writeTimeSeries(tl, fname = "timeseriesdb_export", format = c("csv", "xlsx",
  "json", "rdata"), date_format = NULL, ...)
}
\arguments{
\item{tl}{list of time series}

\item{fname}{character file name. If set to NULL a standard file name chunk + Sys.Date is used.}

\item{date_format}{character denotes the date format. Defaults to NULL. If set to null the default is used: Jan 2010. In combination with LC\_TIME\_Locale various international date formats can be produced.}

\item{wide}{logical. Should the resulting data.frame be cast to wide format? Defaults to TRUE}

\item{xlsx}{logical. Should data be exported to .xlsx? Defaults to FALSE.}

\item{LC_TIME_LOCALE}{character time locale that differs from the standard locale. e.g. en_US.UTF-8. Defaults to NULL and uses the standard locale then.}
}
\description{
TODO: Dox
}
